% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about-main.R
\name{aboutHelperFile}
\alias{aboutHelperFile}
\title{The location of the about module helper file}
\usage{
aboutHelperFile()
}
\value{
string location of the about helper file
}
\description{
The location of the about module helper file
}
\details{
Returns the location of the about helper file
}
\seealso{
Other About:
\code{\link[=aboutServer]{aboutServer()}},
\code{\link[=aboutViewer]{aboutViewer()}}
}
\concept{About}
