% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/about-main.R
\name{aboutViewer}
\alias{aboutViewer}
\title{The module viewer for the shiny app home}
\usage{
aboutViewer(id = "homepage")
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the home page module
}
\description{
The module viewer for the shiny app home
}
\details{
The user specifies the id for the module
}
\seealso{
Other About:
\code{\link[=aboutHelperFile]{aboutHelperFile()}},
\code{\link[=aboutServer]{aboutServer()}}
}
\concept{About}
