% Generated manually to match package documentation style
\name{actionButtonStyleInfo}
\alias{actionButtonStyleInfo}
\title{Action Button Style (Info)}
\usage{
actionButtonStyleInfo()
}
\value{
A CSS style string.
}
\description{
Shared styling for neutral/info action pills used in result tables.
}
