% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-actions.R
\name{actionButtonStyleWarning}
\alias{actionButtonStyleWarning}
\title{Action Button Style (Warning)}
\usage{
actionButtonStyleWarning()
}
\value{
A CSS style string.
}
\description{
Shared styling for warning/attention action pills used in result tables.
}
\seealso{
Other Utils:
\code{\link[=actionButtonStyleInfo]{actionButtonStyleInfo()}},
\code{\link[=createActionButton]{createActionButton()}},
\code{\link[=datasourcesHelperFile]{datasourcesHelperFile()}},
\code{\link[=datasourcesServer]{datasourcesServer()}},
\code{\link[=datasourcesViewer]{datasourcesViewer()}},
\code{\link[=getLogoImage]{getLogoImage()}},
\code{\link[=resultTableViewer]{resultTableViewer()}}
}
\concept{Utils}
