% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-model.R
\name{addValidationName}
\alias{addValidationName}
\title{An internal helper function to add a friendly name for the prediction}
\usage{
addValidationName(result)
}
\arguments{
\item{result}{A data.frame with the prediction performance details}
}
\value{
The result with a new column called friendlyValidationName
}
\description{
This function is not intended for direct use by users.
}
