% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-aggregateFeatures.R
\name{characterizationAggregateFeaturesServer}
\alias{characterizationAggregateFeaturesServer}
\title{The module server for exploring aggregate features results}
\usage{
characterizationAggregateFeaturesServer(
  id,
  connectionHandler,
  mainPanelTab,
  schema,
  tablePrefix,
  cohortTablePrefix = "cg_",
  databaseTable = "DATABASE_META_DATA"
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{mainPanelTab}{the current tab}

\item{schema}{the database schema for the model results}

\item{tablePrefix}{a string that appends the tables in the result schema}

\item{cohortTablePrefix}{a string that appends the COHORT_DEFINITION table in the result schema}

\item{databaseTable}{The database table name}
}
\value{
The server to the description aggregate features module
}
\description{
The module server for exploring aggregate features results
}
\details{
The user specifies the id for the module
}
