% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-aggregateFeatures.R
\name{characterizationAggregateFeaturesViewer}
\alias{characterizationAggregateFeaturesViewer}
\title{The module viewer for exploring aggregate feature results}
\usage{
characterizationAggregateFeaturesViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the description aggregate feature module
}
\description{
The module viewer for exploring aggregate feature results
}
\details{
The user specifies the id for the module
}
