% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-dechallengeRechallenge.R
\name{characterizationDechallengeRechallengeServer}
\alias{characterizationDechallengeRechallengeServer}
\title{The module server for exploring Dechallenge Rechallenge results}
\usage{
characterizationDechallengeRechallengeServer(
  id,
  connectionHandler,
  mainPanelTab,
  schema,
  tablePrefix,
  cohortTablePrefix = "cg_",
  databaseTable = "DATABASE_META_DATA"
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{mainPanelTab}{the current tab}

\item{schema}{the database schema for the model results}

\item{tablePrefix}{a string that appends the tables in the result schema}

\item{cohortTablePrefix}{a string that appends the cohort table in the result schema}

\item{databaseTable}{name of the database table}
}
\value{
The server to the Dechallenge Rechallenge module
}
\description{
The module server for exploring Dechallenge Rechallenge results
}
\details{
The user specifies the id for the module
}
