% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-dechallengeRechallenge.R
\name{characterizationDechallengeRechallengeViewer}
\alias{characterizationDechallengeRechallengeViewer}
\title{The module viewer for exploring Dechallenge Rechallenge results}
\usage{
characterizationDechallengeRechallengeViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the description Dechallenge Rechallenge module
}
\description{
The module viewer for exploring Dechallenge Rechallenge results
}
\details{
The user specifies the id for the module
}
