% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-main.R
\name{characterizationHelperFile}
\alias{characterizationHelperFile}
\title{The location of the characterization module helper file}
\usage{
characterizationHelperFile()
}
\value{
string location of the characterization helper file
}
\description{
The location of the characterization module helper file
}
\details{
Returns the location of the characterization helper file
}
\seealso{
Other Characterization:
\code{\link[=characterizationIncidenceServer]{characterizationIncidenceServer()}},
\code{\link[=characterizationIncidenceViewer]{characterizationIncidenceViewer()}},
\code{\link[=characterizationServer]{characterizationServer()}},
\code{\link[=characterizationViewer]{characterizationViewer()}}
}
\concept{Characterization}
