% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-incidence.R
\name{characterizationIncidenceServer}
\alias{characterizationIncidenceServer}
\title{The module server for exploring incidence results}
\usage{
characterizationIncidenceServer(
  id,
  connectionHandler,
  resultDatabaseSettings,
  reactiveTargetRow,
  reactiveOutcomeTable
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{resultDatabaseSettings}{a list containing the characterization result schema, dbms, tablePrefix, databaseTable and cgTablePrefix}

\item{reactiveTargetRow}{a reactive data.frame with the target of interest details}

\item{reactiveOutcomeTable}{A reactive data.frame with the outcome table for the target of interest}
}
\value{
The server to the prediction incidence module
}
\description{
The module server for exploring incidence results
}
\details{
The user specifies the id for the module
}
\seealso{
Other Characterization:
\code{\link[=characterizationHelperFile]{characterizationHelperFile()}},
\code{\link[=characterizationIncidenceViewer]{characterizationIncidenceViewer()}},
\code{\link[=characterizationServer]{characterizationServer()}},
\code{\link[=characterizationViewer]{characterizationViewer()}}
}
\concept{Characterization}
