% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-incidence.R
\name{characterizationIncidenceViewer}
\alias{characterizationIncidenceViewer}
\title{The module viewer for exploring incidence results}
\usage{
characterizationIncidenceViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the description incidence module
}
\description{
The module viewer for exploring incidence results
}
\details{
The user specifies the id for the module
}
\seealso{
Other Characterization:
\code{\link[=characterizationHelperFile]{characterizationHelperFile()}},
\code{\link[=characterizationIncidenceServer]{characterizationIncidenceServer()}},
\code{\link[=characterizationServer]{characterizationServer()}},
\code{\link[=characterizationViewer]{characterizationViewer()}}
}
\concept{Characterization}
