% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-main.R
\name{characterizationServer}
\alias{characterizationServer}
\title{The module server for exploring characterization studies}
\usage{
characterizationServer(
  id,
  connectionHandler,
  resultDatabaseSettings = list(port = 1)
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list containing the characterization result schema, dbms, tablePrefix, databaseTable and cgTablePrefix}
}
\value{
The server for the characterization module
}
\description{
The module server for exploring characterization studies
}
\details{
The user specifies the id for the module
}
\seealso{
Other Characterization:
\code{\link[=characterizationHelperFile]{characterizationHelperFile()}},
\code{\link[=characterizationIncidenceServer]{characterizationIncidenceServer()}},
\code{\link[=characterizationIncidenceViewer]{characterizationIncidenceViewer()}},
\code{\link[=characterizationViewer]{characterizationViewer()}}
}
\concept{Characterization}
