% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-cohorts.R
\name{characterizationTableServer}
\alias{characterizationTableServer}
\title{The module server for exploring 1 or more cohorts features}
\usage{
characterizationTableServer(
  id,
  connectionHandler,
  mainPanelTab,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{mainPanelTab}{the current tab}

\item{resultDatabaseSettings}{a list containing the characterization result schema, dbms, tablePrefix, databaseTable and cgTablePrefix}
}
\value{
The server to the cohorts features server
}
\description{
The module server for exploring 1 or more cohorts features
}
\details{
The user specifies the id for the module
}
