% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-timeToEvent.R
\name{characterizationTimeToEventServer}
\alias{characterizationTimeToEventServer}
\title{The module server for exploring time to event results}
\usage{
characterizationTimeToEventServer(
  id,
  connectionHandler,
  mainPanelTab,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{mainPanelTab}{the current tab}

\item{resultDatabaseSettings}{a list containing the characterization result schema, dbms, tablePrefix, databaseTable and cgTablePrefix}
}
\value{
The server to the prediction time to event module
}
\description{
The module server for exploring time to event results
}
\details{
The user specifies the id for the module
}
