% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-timeToEvent.R
\name{characterizationTimeToEventViewer}
\alias{characterizationTimeToEventViewer}
\title{The module viewer for exploring time to event results}
\usage{
characterizationTimeToEventViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the characterization time to event module
}
\description{
The module viewer for exploring time to event results
}
\details{
The user specifies the id for the module
}
