% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-characterization.R
\name{characterizationView}
\alias{characterizationView}
\title{characterization}
\usage{
characterizationView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("characterization") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
