% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-counts.R
\name{cohortCountsView}
\alias{cohortCountsView}
\title{Cohort Counts View}
\usage{
cohortCountsView(id)
}
\arguments{
\item{id}{Namespace id}
}
\description{
Shiny view for cohort counts module
}
\seealso{
Other CohortDiagnostics:
\code{\link[=cohortCountsModule]{cohortCountsModule()}},
\code{\link[=cohortDefinitionsModule]{cohortDefinitionsModule()}},
\code{\link[=cohortDefinitionsView]{cohortDefinitionsView()}},
\code{\link[=cohortDiagCharacterizationView]{cohortDiagCharacterizationView()}},
\code{\link[=cohortDiagnosticsHelperFile]{cohortDiagnosticsHelperFile()}},
\code{\link[=cohortDiagnosticsServer]{cohortDiagnosticsServer()}},
\code{\link[=cohortDiagnosticsView]{cohortDiagnosticsView()}},
\code{\link[=cohortOverlapView]{cohortOverlapView()}},
\code{\link[=compareCohortCharacterizationView]{compareCohortCharacterizationView()}},
\code{\link[=conceptsInDataSourceView]{conceptsInDataSourceView()}},
\code{\link[=createCdDatabaseDataSource]{createCdDatabaseDataSource()}},
\code{\link[=databaseInformationView]{databaseInformationView()}},
\code{\link[=getCirceRenderedExpression]{getCirceRenderedExpression()}},
\code{\link[=getEnabledCdReports]{getEnabledCdReports()}},
\code{\link[=incidenceRatesView]{incidenceRatesView()}},
\code{\link[=inclusionRulesView]{inclusionRulesView()}},
\code{\link[=indexEventBreakdownView]{indexEventBreakdownView()}},
\code{\link[=orpahanConceptsView]{orpahanConceptsView()}},
\code{\link[=timeDistributionsView]{timeDistributionsView()}},
\code{\link[=visitContextView]{visitContextView()}}
}
\concept{CohortDiagnostics}
