% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-definition.R
\name{cohortDefinitionsModule}
\alias{cohortDefinitionsModule}
\title{Cohort Definition module}
\usage{
cohortDefinitionsModule(
  id,
  dataSource,
  cohortDefinitions,
  cohortTable = dataSource$cohortTable,
  cohortCountTable = dataSource$cohortCountTable,
  databaseTable = dataSource$dbTable
)
}
\arguments{
\item{id}{Namespace id}

\item{dataSource}{DatabaseConnection}

\item{cohortDefinitions}{reactive of cohort definitions to display}

\item{cohortTable}{data.frame of cohorts, cohortId, cohortName}

\item{cohortCountTable}{data.frame of cohortCounts, cohortId, subjects records}

\item{databaseTable}{data.frame of databasese, databaseId, name}
}
\description{
cohort defintion conceptsets, json etc
}
\seealso{
Other CohortDiagnostics:
\code{\link[=cohortCountsModule]{cohortCountsModule()}},
\code{\link[=cohortCountsView]{cohortCountsView()}},
\code{\link[=cohortDefinitionsView]{cohortDefinitionsView()}},
\code{\link[=cohortDiagCharacterizationView]{cohortDiagCharacterizationView()}},
\code{\link[=cohortDiagnosticsHelperFile]{cohortDiagnosticsHelperFile()}},
\code{\link[=cohortDiagnosticsServer]{cohortDiagnosticsServer()}},
\code{\link[=cohortDiagnosticsView]{cohortDiagnosticsView()}},
\code{\link[=cohortOverlapView]{cohortOverlapView()}},
\code{\link[=compareCohortCharacterizationView]{compareCohortCharacterizationView()}},
\code{\link[=conceptsInDataSourceView]{conceptsInDataSourceView()}},
\code{\link[=createCdDatabaseDataSource]{createCdDatabaseDataSource()}},
\code{\link[=databaseInformationView]{databaseInformationView()}},
\code{\link[=getCirceRenderedExpression]{getCirceRenderedExpression()}},
\code{\link[=getEnabledCdReports]{getEnabledCdReports()}},
\code{\link[=incidenceRatesView]{incidenceRatesView()}},
\code{\link[=inclusionRulesView]{inclusionRulesView()}},
\code{\link[=indexEventBreakdownView]{indexEventBreakdownView()}},
\code{\link[=orpahanConceptsView]{orpahanConceptsView()}},
\code{\link[=timeDistributionsView]{timeDistributionsView()}},
\code{\link[=visitContextView]{visitContextView()}}
}
\concept{CohortDiagnostics}
