% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main.R
\name{cohortDiagnosticsSever}
\alias{cohortDiagnosticsSever}
\title{Cohort Diagnostics Explorer main module}
\usage{
cohortDiagnosticsSever(
  id = "DiagnosticsExplorer",
  connectionHandler = NULL,
  dataSource = NULL,
  resultsDatabaseSettings
)
}
\arguments{
\item{connectionHandler}{ResultModelManager ConnectionHander instance}

\item{resultDatabaseSettings}{results database settings}

\item{vocabularyDatabaseSchema}{Optional vocabulary database schema}
}
\description{
Cohort Diagnostics Explorer main module
}
