% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main.R
\name{cohortDiagnosticsSeverCd}
\alias{cohortDiagnosticsSeverCd}
\title{Cohort Diagnostics Explorer main module}
\usage{
cohortDiagnosticsSeverCd(
  id = "DiagnosticsExplorer",
  connectionHandler,
  resultsDatabaseSettings,
  dataSource
)
}
\arguments{
\item{id}{module Id}

\item{connectionHandler}{ResultModelManager ConnectionHander instance}

\item{dataSource}{cohort diagnostics data source}

\item{resultDatabaseSettings}{results database settings}
}
\description{
Cohort Diagnostics Explorer main module
}
