% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main-ui.R
\name{cohortDiagnosticsView}
\alias{cohortDiagnosticsView}
\title{View for cohort diagnostics module}
\usage{
cohortDiagnosticsView(id = "DiagnosticsExplorer")
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort diagnostics viewer module
}
\description{
View for cohort diagnostics module
}
\details{
The user specifies the id for the module
}
\seealso{
Other CohortDiagnostics:
\code{\link[=cohortCountsModule]{cohortCountsModule()}},
\code{\link[=cohortCountsView]{cohortCountsView()}},
\code{\link[=cohortDefinitionsModule]{cohortDefinitionsModule()}},
\code{\link[=cohortDefinitionsView]{cohortDefinitionsView()}},
\code{\link[=cohortDiagCharacterizationView]{cohortDiagCharacterizationView()}},
\code{\link[=cohortDiagnosticsHelperFile]{cohortDiagnosticsHelperFile()}},
\code{\link[=cohortDiagnosticsServer]{cohortDiagnosticsServer()}},
\code{\link[=cohortOverlapView]{cohortOverlapView()}},
\code{\link[=compareCohortCharacterizationView]{compareCohortCharacterizationView()}},
\code{\link[=conceptsInDataSourceView]{conceptsInDataSourceView()}},
\code{\link[=createCdDatabaseDataSource]{createCdDatabaseDataSource()}},
\code{\link[=databaseInformationView]{databaseInformationView()}},
\code{\link[=getCirceRenderedExpression]{getCirceRenderedExpression()}},
\code{\link[=getEnabledCdReports]{getEnabledCdReports()}},
\code{\link[=incidenceRatesView]{incidenceRatesView()}},
\code{\link[=inclusionRulesView]{inclusionRulesView()}},
\code{\link[=indexEventBreakdownView]{indexEventBreakdownView()}},
\code{\link[=orpahanConceptsView]{orpahanConceptsView()}},
\code{\link[=timeDistributionsView]{timeDistributionsView()}},
\code{\link[=visitContextView]{visitContextView()}}
}
\concept{CohortDiagnostics}
