% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-diagnosticsSummary.R
\name{cohortMethodDiagnosticsSummaryServer}
\alias{cohortMethodDiagnosticsSummaryServer}
\title{The module server for rendering the PLE diagnostics summary}
\usage{
cohortMethodDiagnosticsSummaryServer(
  id,
  connectionHandler,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
the PLE diagnostics summary results
}
\description{
The module server for rendering the PLE diagnostics summary
}
