% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-diagnosticsSummary.R
\name{cohortMethodDiagnosticsSummaryViewer}
\alias{cohortMethodDiagnosticsSummaryViewer}
\title{The module viewer for rendering the PLE diagnostics results}
\usage{
cohortMethodDiagnosticsSummaryViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method diagnostics viewer
}
\description{
The module viewer for rendering the PLE diagnostics results
}
