% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-forestPlot.R
\name{cohortMethodForestPlotViewer}
\alias{cohortMethodForestPlotViewer}
\title{The module viewer for rendering the PLE results forest plot}
\usage{
cohortMethodForestPlotViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method forest plot
}
\description{
The module viewer for rendering the PLE results forest plot
}
