% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-main.R
\name{cohortMethodHelperFile}
\alias{cohortMethodHelperFile}
\title{The location of the cohort method module helper file}
\usage{
cohortMethodHelperFile()
}
\value{
string location of the cohort method helper file
}
\description{
The location of the cohort method module helper file
}
\details{
Returns the location of the cohort method helper file
}
