% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-kaplainMeier.R
\name{cohortMethodKaplanMeierViewer}
\alias{cohortMethodKaplanMeierViewer}
\title{The module viewer for rendering the PLE Kaplan Meier curve}
\usage{
cohortMethodKaplanMeierViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The module viewer for Kaplan Meier objects
}
\description{
The module viewer for rendering the PLE Kaplan Meier curve
}
\seealso{
Other Estimation:
\code{\link[=cohortMethodAttritionServer]{cohortMethodAttritionServer()}},
\code{\link[=cohortMethodAttritionViewer]{cohortMethodAttritionViewer()}},
\code{\link[=cohortMethodCovariateBalanceServer]{cohortMethodCovariateBalanceServer()}},
\code{\link[=cohortMethodCovariateBalanceViewer]{cohortMethodCovariateBalanceViewer()}},
\code{\link[=cohortMethodKaplanMeierServer]{cohortMethodKaplanMeierServer()}},
\code{\link[=cohortMethodPopulationCharacteristicsServer]{cohortMethodPopulationCharacteristicsServer()}},
\code{\link[=cohortMethodPopulationCharacteristicsViewer]{cohortMethodPopulationCharacteristicsViewer()}},
\code{\link[=cohortMethodPowerServer]{cohortMethodPowerServer()}},
\code{\link[=cohortMethodPowerViewer]{cohortMethodPowerViewer()}},
\code{\link[=cohortMethodPropensityModelServer]{cohortMethodPropensityModelServer()}},
\code{\link[=cohortMethodPropensityModelViewer]{cohortMethodPropensityModelViewer()}},
\code{\link[=cohortMethodPropensityScoreDistServer]{cohortMethodPropensityScoreDistServer()}},
\code{\link[=cohortMethodPropensityScoreDistViewer]{cohortMethodPropensityScoreDistViewer()}},
\code{\link[=cohortMethodSystematicErrorServer]{cohortMethodSystematicErrorServer()}},
\code{\link[=cohortMethodSystematicErrorViewer]{cohortMethodSystematicErrorViewer()}},
\code{\link[=estimationHelperFile]{estimationHelperFile()}},
\code{\link[=estimationServer]{estimationServer()}},
\code{\link[=estimationViewer]{estimationViewer()}}
}
\concept{Estimation}
