% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-propensityModel.R
\name{cohortMethodPropensityModelViewer}
\alias{cohortMethodPropensityModelViewer}
\title{The module viewer for rendering the PLE propensity score model covariates/coefficients}
\usage{
cohortMethodPropensityModelViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method propensity score model covariates/coefficients
}
\description{
The module viewer for rendering the PLE propensity score model covariates/coefficients
}
\seealso{
Other Estimation:
\code{\link[=cohortMethodAttritionServer]{cohortMethodAttritionServer()}},
\code{\link[=cohortMethodAttritionViewer]{cohortMethodAttritionViewer()}},
\code{\link[=cohortMethodCovariateBalanceServer]{cohortMethodCovariateBalanceServer()}},
\code{\link[=cohortMethodCovariateBalanceViewer]{cohortMethodCovariateBalanceViewer()}},
\code{\link[=cohortMethodKaplanMeierServer]{cohortMethodKaplanMeierServer()}},
\code{\link[=cohortMethodKaplanMeierViewer]{cohortMethodKaplanMeierViewer()}},
\code{\link[=cohortMethodPopulationCharacteristicsServer]{cohortMethodPopulationCharacteristicsServer()}},
\code{\link[=cohortMethodPopulationCharacteristicsViewer]{cohortMethodPopulationCharacteristicsViewer()}},
\code{\link[=cohortMethodPowerServer]{cohortMethodPowerServer()}},
\code{\link[=cohortMethodPowerViewer]{cohortMethodPowerViewer()}},
\code{\link[=cohortMethodPropensityModelServer]{cohortMethodPropensityModelServer()}},
\code{\link[=cohortMethodPropensityScoreDistServer]{cohortMethodPropensityScoreDistServer()}},
\code{\link[=cohortMethodPropensityScoreDistViewer]{cohortMethodPropensityScoreDistViewer()}},
\code{\link[=cohortMethodSystematicErrorServer]{cohortMethodSystematicErrorServer()}},
\code{\link[=cohortMethodSystematicErrorViewer]{cohortMethodSystematicErrorViewer()}},
\code{\link[=estimationHelperFile]{estimationHelperFile()}},
\code{\link[=estimationServer]{estimationServer()}},
\code{\link[=estimationViewer]{estimationViewer()}}
}
\concept{Estimation}
