% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/estimation-cohort-method-propensityScoreDistribution.R
\name{cohortMethodPropensityScoreDistServer}
\alias{cohortMethodPropensityScoreDistServer}
\title{The module server for rendering a PLE propensity score distribution}
\usage{
cohortMethodPropensityScoreDistServer(
  id,
  selectedRow,
  connectionHandler,
  resultDatabaseSettings,
  metaAnalysisDbIds = F
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}

\item{metaAnalysisDbIds}{metaAnalysisDbIds}
}
\value{
the PLE propensity score distribution content server
}
\description{
The module server for rendering a PLE propensity score distribution
}
\seealso{
Other Estimation:
\code{\link[=cohortMethodAttritionServer]{cohortMethodAttritionServer()}},
\code{\link[=cohortMethodAttritionViewer]{cohortMethodAttritionViewer()}},
\code{\link[=cohortMethodCovariateBalanceServer]{cohortMethodCovariateBalanceServer()}},
\code{\link[=cohortMethodCovariateBalanceViewer]{cohortMethodCovariateBalanceViewer()}},
\code{\link[=cohortMethodKaplanMeierServer]{cohortMethodKaplanMeierServer()}},
\code{\link[=cohortMethodKaplanMeierViewer]{cohortMethodKaplanMeierViewer()}},
\code{\link[=cohortMethodPopulationCharacteristicsServer]{cohortMethodPopulationCharacteristicsServer()}},
\code{\link[=cohortMethodPopulationCharacteristicsViewer]{cohortMethodPopulationCharacteristicsViewer()}},
\code{\link[=cohortMethodPowerServer]{cohortMethodPowerServer()}},
\code{\link[=cohortMethodPowerViewer]{cohortMethodPowerViewer()}},
\code{\link[=cohortMethodPropensityModelServer]{cohortMethodPropensityModelServer()}},
\code{\link[=cohortMethodPropensityModelViewer]{cohortMethodPropensityModelViewer()}},
\code{\link[=cohortMethodPropensityScoreDistViewer]{cohortMethodPropensityScoreDistViewer()}},
\code{\link[=cohortMethodSystematicErrorServer]{cohortMethodSystematicErrorServer()}},
\code{\link[=cohortMethodSystematicErrorViewer]{cohortMethodSystematicErrorViewer()}},
\code{\link[=estimationHelperFile]{estimationHelperFile()}},
\code{\link[=estimationServer]{estimationServer()}},
\code{\link[=estimationViewer]{estimationViewer()}}
}
\concept{Estimation}
