% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/estimation-cohort-method-propensityScoreDistribution.R
\name{cohortMethodPropensityScoreDistViewer}
\alias{cohortMethodPropensityScoreDistViewer}
\title{The module viewer for rendering the propensity score distribution}
\usage{
cohortMethodPropensityScoreDistViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method propensity score distribution
}
\description{
The module viewer for rendering the propensity score distribution
}
\seealso{
Other Estimation:
\code{\link[=cohortMethodAttritionServer]{cohortMethodAttritionServer()}},
\code{\link[=cohortMethodAttritionViewer]{cohortMethodAttritionViewer()}},
\code{\link[=cohortMethodCovariateBalanceServer]{cohortMethodCovariateBalanceServer()}},
\code{\link[=cohortMethodCovariateBalanceViewer]{cohortMethodCovariateBalanceViewer()}},
\code{\link[=cohortMethodKaplanMeierServer]{cohortMethodKaplanMeierServer()}},
\code{\link[=cohortMethodKaplanMeierViewer]{cohortMethodKaplanMeierViewer()}},
\code{\link[=cohortMethodPopulationCharacteristicsServer]{cohortMethodPopulationCharacteristicsServer()}},
\code{\link[=cohortMethodPopulationCharacteristicsViewer]{cohortMethodPopulationCharacteristicsViewer()}},
\code{\link[=cohortMethodPowerServer]{cohortMethodPowerServer()}},
\code{\link[=cohortMethodPowerViewer]{cohortMethodPowerViewer()}},
\code{\link[=cohortMethodPropensityModelServer]{cohortMethodPropensityModelServer()}},
\code{\link[=cohortMethodPropensityModelViewer]{cohortMethodPropensityModelViewer()}},
\code{\link[=cohortMethodPropensityScoreDistServer]{cohortMethodPropensityScoreDistServer()}},
\code{\link[=cohortMethodSystematicErrorServer]{cohortMethodSystematicErrorServer()}},
\code{\link[=cohortMethodSystematicErrorViewer]{cohortMethodSystematicErrorViewer()}},
\code{\link[=estimationHelperFile]{estimationHelperFile()}},
\code{\link[=estimationServer]{estimationServer()}},
\code{\link[=estimationViewer]{estimationViewer()}}
}
\concept{Estimation}
