% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-resultsTable.R
\name{cohortMethodResultsTableServer}
\alias{cohortMethodResultsTableServer}
\title{The module server for rendering the PLE results per current selections}
\usage{
cohortMethodResultsTableServer(
  id,
  connectionHandler,
  inputParams,
  resultsSchema,
  tablePrefix,
  databaseTable
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the PLE results database}

\item{inputParams}{the selected study parameters of interest}

\item{resultsSchema}{the schema with the PLE results}

\item{tablePrefix}{tablePrefix}

\item{databaseTable}{databaseTable}
}
\value{
the PLE main results table server server
}
\description{
The module server for rendering the PLE results per current selections
}
