% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-resultsTable.R
\name{cohortMethodResultsTableViewer}
\alias{cohortMethodResultsTableViewer}
\title{The module viewer for rendering the PLE main results}
\usage{
cohortMethodResultsTableViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the PLE main results
}
\description{
The module viewer for rendering the PLE main results
}
