% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-resultSummary.R
\name{cohortMethodResultSummaryServer}
\alias{cohortMethodResultSummaryServer}
\title{The module server for rendering the PLE diagnostics summary}
\usage{
cohortMethodResultSummaryServer(
  id,
  connectionHandler,
  resultDatabaseSettings,
  inputSelected
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}

\item{inputSelected}{The target id, comparator id, outcome id and analysis id selected by the user}
}
\value{
the PLE diagnostics summary results
}
\description{
The module server for rendering the PLE diagnostics summary
}
