% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-resultSummary.R
\name{cohortMethodResultSummaryViewer}
\alias{cohortMethodResultSummaryViewer}
\title{The module viewer for rendering the cohort method results}
\usage{
cohortMethodResultSummaryViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method diagnostics viewer
}
\description{
The module viewer for rendering the cohort method results
}
