% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-main.R
\name{cohortMethodServer}
\alias{cohortMethodServer}
\title{The module server for the main cohort method module}
\usage{
cohortMethodServer(id, connectionHandler, resultDatabaseSettings)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a named list containing the PLE results database connection details}
}
\value{
the PLE results viewer main module server
}
\description{
The module server for the main cohort method module
}
