% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-subgroups.R
\name{cohortMethodSubgroupsViewer}
\alias{cohortMethodSubgroupsViewer}
\title{The module viewer for rendering the PLE subgroup results}
\usage{
cohortMethodSubgroupsViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method subgroup results module
}
\description{
The module viewer for rendering the PLE subgroup results
}
