% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-main.R
\name{cohortMethodViewer}
\alias{cohortMethodViewer}
\title{The viewer of the main cohort method module}
\usage{
cohortMethodViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method results viewer
}
\description{
The viewer of the main cohort method module
}
