% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-cohort-overlap.R
\name{cohortOverlapModule}
\alias{cohortOverlapModule}
\title{Cohort Overlap Module}
\usage{
cohortOverlapModule(
  id,
  dataSource,
  selectedCohorts,
  selectedDatabaseIds,
  targetCohortId,
  cohortIds,
  cohortTable
)
}
\description{
Cohort Overlap Module
}
