% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortModule.R
\name{cohortServer}
\alias{cohortServer}
\title{The server for extracting cohorts from an ATLAS webapi}
\usage{
cohortServer(input, output, session, cohortReactive)
}
\arguments{
\item{input}{Standard for shiny modules}

\item{output}{Standard for shiny modules}

\item{session}{Standard for shiny modules}

\item{cohortReactive}{The list of cohorts from the webApi (extracted using the extractCohortsServer module)}
}
\description{
The server for extracting cohorts from an ATLAS webapi
}
\examples{
\dontrun{
cohortReactive <- callModule(extractCohortsServer, 'cohortExtract', webApi = webApi)
targetList <- callModule(cohortServer, 'cohort1', cohortReactive)
}
}
