% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortModule.R
\name{cohortViewer}
\alias{cohortViewer}
\title{The user interface for viewing and extracting cohorts from an ATLAS webapi}
\usage{
cohortViewer(id, labelv = "cohorts")
}
\arguments{
\item{id}{The module identifier}

\item{labelv}{A string that is the header for the box containing the dropdown selector that lets you pick the cohort of interest into a list}
}
\description{
The user interface for viewing and extracting cohorts from an ATLAS webapi
}
\examples{
\dontrun{
cohortViewer('cohort1', 'Target Cohort')
}
}
