% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-compareCharacterization.R
\name{compareCohortCharacterizationView}
\alias{compareCohortCharacterizationView}
\title{compare characterization view}
\usage{
compareCohortCharacterizationView(
  id,
  title = "Compare cohort characterization"
)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("compareCohortCharacterization") inside diagnosticsExplorer module}

\item{title}{Optional string title field}
}
\description{
Use for customizing UI
}
\seealso{
Other CohortDiagnostics:
\code{\link[=cohortCountsModule]{cohortCountsModule()}},
\code{\link[=cohortCountsView]{cohortCountsView()}},
\code{\link[=cohortDefinitionsModule]{cohortDefinitionsModule()}},
\code{\link[=cohortDefinitionsView]{cohortDefinitionsView()}},
\code{\link[=cohortDiagCharacterizationView]{cohortDiagCharacterizationView()}},
\code{\link[=cohortDiagnosticsHelperFile]{cohortDiagnosticsHelperFile()}},
\code{\link[=cohortDiagnosticsServer]{cohortDiagnosticsServer()}},
\code{\link[=cohortDiagnosticsView]{cohortDiagnosticsView()}},
\code{\link[=cohortOverlapView]{cohortOverlapView()}},
\code{\link[=conceptsInDataSourceView]{conceptsInDataSourceView()}},
\code{\link[=createCdDatabaseDataSource]{createCdDatabaseDataSource()}},
\code{\link[=databaseInformationView]{databaseInformationView()}},
\code{\link[=getCirceRenderedExpression]{getCirceRenderedExpression()}},
\code{\link[=getEnabledCdReports]{getEnabledCdReports()}},
\code{\link[=incidenceRatesView]{incidenceRatesView()}},
\code{\link[=inclusionRulesView]{inclusionRulesView()}},
\code{\link[=indexEventBreakdownView]{indexEventBreakdownView()}},
\code{\link[=orpahanConceptsView]{orpahanConceptsView()}},
\code{\link[=timeDistributionsView]{timeDistributionsView()}},
\code{\link[=visitContextView]{visitContextView()}}
}
\concept{CohortDiagnostics}
