% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main-ui.R
\name{conceptsInDataSourceUi}
\alias{conceptsInDataSourceUi}
\title{The user interface to the cohort diagnostics viewer module}
\usage{
conceptsInDataSourceUi(id, parentId = "DiagnosticsExplorer")
}
\arguments{
\item{id}{id}

\item{parentId}{parent namespace}
}
\description{
The user interface to the cohort diagnostics viewer module
}
