% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateModule.R
\name{covariateViewer}
\alias{covariateViewer}
\title{The user interface for creating covariate settings for FeatureExtraction}
\usage{
covariateViewer(id, labelv = "Covariates")
}
\arguments{
\item{id}{The module identifier}

\item{labelv}{A string label used for the form box}
}
\description{
The user interface for creating covariate settings for FeatureExtraction
}
\examples{
\dontrun{
covariateViewer('covariate1', labelv = "Covariates")
}
}
