% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heplers-componentsCreateCustomColDefList.R
\name{createCustomColDefList}
\alias{createCustomColDefList}
\title{Creating a list of custom column definitions for use in reactables}
\usage{
createCustomColDefList(rawColNames, niceColNames, tooltipText)
}
\arguments{
\item{rawColNames}{The raw column names taken directly from the source
data table that are to be overwritten in the reactable}

\item{niceColNames}{The formatted column names that will appear as-specified in 
the reactable}

\item{tooltipText}{The text to be displayed in a toolTip when hovering over the 
column in the reactable}
}
\value{
A named list of reactable::colDef objects
}
\description{
Creating a list of custom column definitions for use in reactables
}
\examples{
createCustomColDefList(rawColNames = c("firstName", "lastName"), niceColNames = c("First Name", "Last Name"), tooltipText = c("The person's first name", "The person's last name"))

}
