% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-diagnostic-main.R
\name{dataDiagnosticHelperFile}
\alias{dataDiagnosticHelperFile}
\title{The location of the data-diagnostic module helper file}
\usage{
dataDiagnosticHelperFile()
}
\value{
string location of the data-diagnostic helper file
}
\description{
The location of the data-diagnostic module helper file
}
\details{
Returns the location of the data-diagnostic helper file
}
\seealso{
Other DataDiagnostics:
\code{\link[=dataDiagnosticDrillServer]{dataDiagnosticDrillServer()}},
\code{\link[=dataDiagnosticDrillViewer]{dataDiagnosticDrillViewer()}},
\code{\link[=dataDiagnosticServer]{dataDiagnosticServer()}},
\code{\link[=dataDiagnosticSummaryServer]{dataDiagnosticSummaryServer()}},
\code{\link[=dataDiagnosticSummaryViewer]{dataDiagnosticSummaryViewer()}},
\code{\link[=dataDiagnosticViewer]{dataDiagnosticViewer()}}
}
\concept{DataDiagnostics}
