% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-main.R
\name{descriptionHelperFile}
\alias{descriptionHelperFile}
\title{The location of the description module helper file}
\usage{
descriptionHelperFile()
}
\value{
string location of the description helper file
}
\description{
The location of the description module helper file
}
\details{
Returns the location of the description helper file
}
