% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-incidence.R
\name{descriptionIncidenceServer}
\alias{descriptionIncidenceServer}
\title{The module server for exploring incidence results}
\usage{
descriptionIncidenceServer(
  id,
  con,
  mainPanelTab,
  schema,
  dbms,
  incidenceTablePrefix,
  tempEmulationSchema = NULL,
  databaseTable = "DATABASE_META_DATA"
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{con}{the connection to the prediction result database}

\item{mainPanelTab}{the current tab}

\item{schema}{the database schema for the model results}

\item{dbms}{the database management system for the model results}

\item{incidenceTablePrefix}{a string that appends the incidence table in the result schema}

\item{tempEmulationSchema}{The temp schema (optional)}

\item{databaseTable}{name of the database table}
}
\value{
The server to the prediction incidence module
}
\description{
The module server for exploring incidence results
}
\details{
The user specifies the id for the module
}
