% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-incidence.R
\name{descriptionIncidenceViewer}
\alias{descriptionIncidenceViewer}
\title{The module viewer for exploring incidence results}
\usage{
descriptionIncidenceViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the description incidence module
}
\description{
The module viewer for exploring incidence results
}
\details{
The user specifies the id for the module
+
}
