% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-cohorts.R
\name{descriptionTableServer}
\alias{descriptionTableServer}
\title{The module server for exploring 1 or more cohorts features}
\usage{
descriptionTableServer(
  id,
  connectionHandler,
  mainPanelTab,
  schema,
  tablePrefix,
  cohortTablePrefix,
  databaseTable = "DATABASE_META_DATA"
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{mainPanelTab}{the current tab}

\item{schema}{the database schema for the model results}

\item{tablePrefix}{a string that appends the tables in the result schema}

\item{cohortTablePrefix}{a string that appends the cohort table in the result schema}

\item{databaseTable}{name of the database table}
}
\value{
The server to the cohorts features server
}
\description{
The module server for exploring 1 or more cohorts features
}
\details{
The user specifies the id for the module
}
