% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-cohorts.R
\name{descriptionTableViewer}
\alias{descriptionTableViewer}
\title{The module viewer for exploring 1 or more cohorts features}
\usage{
descriptionTableViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the description cohorts features
}
\description{
The module viewer for exploring 1 or more cohorts features
}
\details{
The user specifies the id for the module
}
