% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-timeToEvent.R
\name{descriptionTimeToEventServer}
\alias{descriptionTimeToEventServer}
\title{The module server for exploring time to event results}
\usage{
descriptionTimeToEventServer(
  id,
  con,
  mainPanelTab,
  schema,
  dbms,
  tablePrefix,
  tempEmulationSchema
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{con}{the connection to the prediction result database}

\item{mainPanelTab}{the current tab}

\item{schema}{the database schema for the model results}

\item{dbms}{the database management system for the model results}

\item{tablePrefix}{a string that appends the tables in the result schema}

\item{tempEmulationSchema}{The temp schema (optional)}
}
\value{
The server to the prediction time to event module
}
\description{
The module server for exploring time to event results
}
\details{
The user specifies the id for the module
}
