% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-timeToEvent.R
\name{descriptionTimeToEventViewer}
\alias{descriptionTimeToEventViewer}
\title{The module viewer for exploring time to event results}
\usage{
descriptionTimeToEventViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the description time to event module
}
\description{
The module viewer for exploring time to event results
}
\details{
The user specifies the id for the module
}
