% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-main.R
\name{descriptionViewer}
\alias{descriptionViewer}
\title{The module viewer for exploring description studies}
\usage{
descriptionViewer(id = 1)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the description viewer module
}
\description{
The module viewer for exploring description studies
}
\details{
The user specifies the id for the module
}
