% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-attrition.R
\name{estimationAttritionServer}
\alias{estimationAttritionServer}
\title{The module server for rendering the PLE attrition results}
\usage{
estimationAttritionServer(
  id,
  selectedRow,
  inputParams,
  connection,
  resultsSchema,
  tablePrefix,
  databaseTable
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}

\item{connection}{the connection to the PLE results database}

\item{resultsSchema}{the schema with the PLE results}
}
\value{
the PLE attrition results content server
}
\description{
The module server for rendering the PLE attrition results
}
