% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-attrition.R
\name{estimationAttritionViewer}
\alias{estimationAttritionViewer}
\title{The module viewer for rendering the PLE attrition results}
\usage{
estimationAttritionViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation attrition
}
\description{
The module viewer for rendering the PLE attrition results
}
