% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-covariateBalance.R
\name{estimationCovariateBalanceViewer}
\alias{estimationCovariateBalanceViewer}
\title{The module viewer for rendering the PLE covariate balance analysis}
\usage{
estimationCovariateBalanceViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation covariate balance results
}
\description{
The module viewer for rendering the PLE covariate balance analysis
}
