% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-diagnosticsSummary.R
\name{estimationDiagnosticsSummaryServer}
\alias{estimationDiagnosticsSummaryServer}
\title{The module server for rendering the PLE diagnostics summary}
\usage{
estimationDiagnosticsSummaryServer(id, connection, resultsSchema)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connection}{the connection to the PLE results database}

\item{resultsSchema}{the schema with the PLE results}
}
\value{
the PLE diagnostics summary results
}
\description{
The module server for rendering the PLE diagnostics summary
}
