% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-diagnosticsSummary.R
\name{estimationDiagnosticsSummaryViewer}
\alias{estimationDiagnosticsSummaryViewer}
\title{The module viewer for rendering the PLE diagnostics results}
\usage{
estimationDiagnosticsSummaryViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation diagnostics viewer
}
\description{
The module viewer for rendering the PLE diagnostics results
}
