% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-forestPlot.R
\name{estimationForestPlotServer}
\alias{estimationForestPlotServer}
\title{The module server for rendering the PLE multiple results forest plot}
\usage{
estimationForestPlotServer(id, selectedRow, inputParams)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}
}
\value{
the PLE forest plot content server
}
\description{
The module server for rendering the PLE multiple results forest plot
}
