% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-forestPlot.R
\name{estimationForestPlotViewer}
\alias{estimationForestPlotViewer}
\title{The module viewer for rendering the PLE results forest plot}
\usage{
estimationForestPlotViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation forest plot
}
\description{
The module viewer for rendering the PLE results forest plot
}
