% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-kaplainMeier.R
\name{estimationKaplanMeierViewer}
\alias{estimationKaplanMeierViewer}
\title{The module viewer for rendering the PLE Kaplan Meier curve}
\usage{
estimationKaplanMeierViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The module viewer for Kaplan Meier objects
}
\description{
The module viewer for rendering the PLE Kaplan Meier curve
}
