% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-populationCharacteristics.R
\name{estimationPopulationCharacteristicsServer}
\alias{estimationPopulationCharacteristicsServer}
\title{The module server for rendering the population characteristics}
\usage{
estimationPopulationCharacteristicsServer(
  id,
  selectedRow,
  inputParams,
  connectionHandler,
  resultsSchema,
  tablePrefix
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultsSchema}{the schema with the PLE results}

\item{tablePrefix}{tablePrefix}
}
\value{
the PLE population characteristics content server
}
\description{
The module server for rendering the population characteristics
}
