% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-populationCharacteristics.R
\name{estimationPopulationCharacteristicsViewer}
\alias{estimationPopulationCharacteristicsViewer}
\title{The module viewer for rendering the PLE population characteristics}
\usage{
estimationPopulationCharacteristicsViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation population characteristics objects
}
\description{
The module viewer for rendering the PLE population characteristics
}
