% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-power.R
\name{estimationPowerViewer}
\alias{estimationPowerViewer}
\title{The module viewer for rendering the PLE power analysis}
\usage{
estimationPowerViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation power calculation results
}
\description{
The module viewer for rendering the PLE power analysis
}
