% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-propensityModel.R
\name{estimationPropensityModelServer}
\alias{estimationPropensityModelServer}
\title{The module server for rendering the propensity score model}
\usage{
estimationPropensityModelServer(
  id,
  selectedRow,
  inputParams,
  connection,
  resultsSchema,
  tablePrefix
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}

\item{connection}{the connection to the PLE results database}

\item{resultsSchema}{the schema with the PLE results}
}
\value{
the PLE propensity score model
}
\description{
The module server for rendering the propensity score model
}
