% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-propensityModel.R
\name{estimationPropensityModelViewer}
\alias{estimationPropensityModelViewer}
\title{The module viewer for rendering the PLE propensity score model covariates/coefficients}
\usage{
estimationPropensityModelViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation propensity score model covariates/coefficients
}
\description{
The module viewer for rendering the PLE propensity score model covariates/coefficients
}
