% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-propensityScoreDistribution.R
\name{estimationPropensityScoreDistViewer}
\alias{estimationPropensityScoreDistViewer}
\title{The module viewer for rendering the propensity score distribution}
\usage{
estimationPropensityScoreDistViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation propensity score distribution
}
\description{
The module viewer for rendering the propensity score distribution
}
