% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-resultsTable.R
\name{estimationResultsTableServer}
\alias{estimationResultsTableServer}
\title{The module server for rendering the PLE results per current selections}
\usage{
estimationResultsTableServer(
  id,
  connection,
  inputParams,
  resultsSchema,
  tablePrefix,
  databaseTable
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connection}{the connection to the PLE results database}

\item{inputParams}{the selected study parameters of interest}

\item{resultsSchema}{the schema with the PLE results}
}
\value{
the PLE main results table server server
}
\description{
The module server for rendering the PLE results per current selections
}
