% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-resultsTable.R
\name{estimationResultsTableViewer}
\alias{estimationResultsTableViewer}
\title{The module viewer for rendering the PLE main results}
\usage{
estimationResultsTableViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the PLE main results
}
\description{
The module viewer for rendering the PLE main results
}
