% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-main.R
\name{estimationServer}
\alias{estimationServer}
\title{The module server for exploring estimation studies}
\usage{
estimationServer(
  id,
  connectionHandler,
  resultDatabaseSettings = list(port = 1)
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list containing the characterization result schema, dbms, tablePrefix, databaseTable and cgTablePrefix}
}
\value{
The server for the estimation module
}
\description{
The module server for exploring estimation studies
}
\details{
The user specifies the id for the module
}
\seealso{
Other Estimation:
\code{\link[=cohortMethodAttritionServer]{cohortMethodAttritionServer()}},
\code{\link[=cohortMethodAttritionViewer]{cohortMethodAttritionViewer()}},
\code{\link[=cohortMethodCovariateBalanceServer]{cohortMethodCovariateBalanceServer()}},
\code{\link[=cohortMethodCovariateBalanceViewer]{cohortMethodCovariateBalanceViewer()}},
\code{\link[=cohortMethodKaplanMeierServer]{cohortMethodKaplanMeierServer()}},
\code{\link[=cohortMethodKaplanMeierViewer]{cohortMethodKaplanMeierViewer()}},
\code{\link[=cohortMethodPopulationCharacteristicsServer]{cohortMethodPopulationCharacteristicsServer()}},
\code{\link[=cohortMethodPopulationCharacteristicsViewer]{cohortMethodPopulationCharacteristicsViewer()}},
\code{\link[=cohortMethodPowerServer]{cohortMethodPowerServer()}},
\code{\link[=cohortMethodPowerViewer]{cohortMethodPowerViewer()}},
\code{\link[=cohortMethodPropensityModelServer]{cohortMethodPropensityModelServer()}},
\code{\link[=cohortMethodPropensityModelViewer]{cohortMethodPropensityModelViewer()}},
\code{\link[=cohortMethodPropensityScoreDistServer]{cohortMethodPropensityScoreDistServer()}},
\code{\link[=cohortMethodPropensityScoreDistViewer]{cohortMethodPropensityScoreDistViewer()}},
\code{\link[=cohortMethodSystematicErrorServer]{cohortMethodSystematicErrorServer()}},
\code{\link[=cohortMethodSystematicErrorViewer]{cohortMethodSystematicErrorViewer()}},
\code{\link[=estimationHelperFile]{estimationHelperFile()}},
\code{\link[=estimationViewer]{estimationViewer()}}
}
\concept{Estimation}
