% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-subgroups.R
\name{estimationSubgroupsServer}
\alias{estimationSubgroupsServer}
\title{The module server for rendering the subgroup results}
\usage{
estimationSubgroupsServer(id, selectedRow, inputParams)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}
}
\value{
the PLE subgroup results server
}
\description{
The module server for rendering the subgroup results
}
