% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-subgroups.R
\name{estimationSubgroupsViewer}
\alias{estimationSubgroupsViewer}
\title{The module viewer for rendering the PLE subgroup results}
\usage{
estimationSubgroupsViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation subgroup results module
}
\description{
The module viewer for rendering the PLE subgroup results
}
