% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-systematicError.R
\name{estimationSystematicErrorViewer}
\alias{estimationSystematicErrorViewer}
\title{The module viewer for rendering the PLE systematic error objects}
\usage{
estimationSystematicErrorViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation systematic error module
}
\description{
The module viewer for rendering the PLE systematic error objects
}
