% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-titlePanel.R
\name{estimationTitlePanelServer}
\alias{estimationTitlePanelServer}
\title{The module server for rendering the estimation title}
\usage{
estimationTitlePanelServer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The server for the estimation title server
}
\description{
The module server for rendering the estimation title
}
