% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-titlePanel.R
\name{estimationTitlePanelViewer}
\alias{estimationTitlePanelViewer}
\title{The module viewer for rendering the estimation title}
\usage{
estimationTitlePanelViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation title module
}
\description{
The module viewer for rendering the estimation title
}
