% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-synth-main.R
\name{evidenceSynthesisHelperFile}
\alias{evidenceSynthesisHelperFile}
\title{The location of the evidence synthesis module helper file}
\usage{
evidenceSynthesisHelperFile()
}
\value{
string location of the evidence synthesis helper file
}
\description{
The location of the evidence synthesis module helper file
}
\details{
Returns the location of the evidence synthesis helper file
}
