% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-synth-main.R
\name{evidenceSynthesisServer}
\alias{evidenceSynthesisServer}
\title{The module server for exploring PatientLevelPrediction}
\usage{
evidenceSynthesisServer(
  id,
  connectionHandler,
  resultDatabaseSettings = list(port = 1)
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
The server for the PatientLevelPrediction module
}
\description{
The module server for exploring PatientLevelPrediction
}
\details{
The user specifies the id for the module
}
