% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evidence-synth-main.R
\name{evidenceSynthesisViewer}
\alias{evidenceSynthesisViewer}
\title{The module viewer for exploring evidence-synthesis}
\usage{
evidenceSynthesisViewer(id = 1)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the evidence-synthesis viewer module
}
\description{
The module viewer for exploring evidence-synthesis
}
\details{
The user specifies the id for the module
}
