% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortExtractModule.R
\name{extractCohortsServer}
\alias{extractCohortsServer}
\title{The server for extracting cohorts from an ATLAS webapi}
\usage{
extractCohortsServer(input, output, session, webApi)
}
\arguments{
\item{input}{Standard for shiny modules}

\item{output}{Standard for shiny modules}

\item{session}{Standard for shiny modules}

\item{webApi}{The webApi locations (output from the webApiServer module)}
}
\description{
The server for extracting cohorts from an ATLAS webapi
}
\examples{
\dontrun{
webApi <- callModule(webApiServer, 'webApiMain')
cohortReactive <- callModule(extractCohortsServer, 'extractCohort1',
                            webApi = webApi)
}
}
