% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortExtractModule.R
\name{extractCohortsViewer}
\alias{extractCohortsViewer}
\title{The user interface for extracting cohorts from an ATLAS webapi}
\usage{
extractCohortsViewer(id)
}
\arguments{
\item{id}{The module identifier}
}
\description{
The user interface for extracting cohorts from an ATLAS webapi
}
\examples{
\dontrun{
extractCohortsViewer('extractCohort1')
}
}
