% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-definition.R
\name{getCirceRenderedExpression}
\alias{getCirceRenderedExpression}
\title{Returns list with circe generated documentation}
\usage{
getCirceRenderedExpression(
  cohortDefinition,
  cohortName = "Cohort Definition",
  includeConceptSets = FALSE
)
}
\arguments{
\item{cohortDefinition}{An R object (list) with a list representation of the cohort definition expression,
that may be converted to a cohort expression JSON using
jsonlite::toJSON(x = cohortDefinition, digits = 23, pretty = TRUE)}

\item{cohortName}{Name for the cohort definition}

\item{includeConceptSets}{Do you want to inclued concept set in the documentation}
}
\value{
list object
}
\description{
Returns list with circe generated documentation
}
\seealso{
Other CohortDiagnostics:
\code{\link[=cohortCountsModule]{cohortCountsModule()}},
\code{\link[=cohortCountsView]{cohortCountsView()}},
\code{\link[=cohortDefinitionsModule]{cohortDefinitionsModule()}},
\code{\link[=cohortDefinitionsView]{cohortDefinitionsView()}},
\code{\link[=cohortDiagCharacterizationView]{cohortDiagCharacterizationView()}},
\code{\link[=cohortDiagnosticsHelperFile]{cohortDiagnosticsHelperFile()}},
\code{\link[=cohortDiagnosticsServer]{cohortDiagnosticsServer()}},
\code{\link[=cohortDiagnosticsView]{cohortDiagnosticsView()}},
\code{\link[=cohortOverlapView]{cohortOverlapView()}},
\code{\link[=compareCohortCharacterizationView]{compareCohortCharacterizationView()}},
\code{\link[=conceptsInDataSourceView]{conceptsInDataSourceView()}},
\code{\link[=createCdDatabaseDataSource]{createCdDatabaseDataSource()}},
\code{\link[=databaseInformationView]{databaseInformationView()}},
\code{\link[=getEnabledCdReports]{getEnabledCdReports()}},
\code{\link[=incidenceRatesView]{incidenceRatesView()}},
\code{\link[=inclusionRulesView]{inclusionRulesView()}},
\code{\link[=indexEventBreakdownView]{indexEventBreakdownView()}},
\code{\link[=orpahanConceptsView]{orpahanConceptsView()}},
\code{\link[=timeDistributionsView]{timeDistributionsView()}},
\code{\link[=visitContextView]{visitContextView()}}
}
\concept{CohortDiagnostics}
