% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-compareCharacterization.R
\name{getCohortRelationshipCharacterizationResults}
\alias{getCohortRelationshipCharacterizationResults}
\title{Returns cohort as feature characterization}
\usage{
getCohortRelationshipCharacterizationResults(
  dataSource,
  cohortIds,
  databaseIds
)
}
\value{
Returns a list object with covariateValue,
covariateRef, analysisRef output of cohort as features. To avoid clash
with covaraiteId and conceptId returned from Feature Extraction
the output is a negative integer.
}
\description{
Returns a list object with covariateValue,
covariateRef, analysisRef output of cohort as features.

template DataSource

template CohortIds

template DatabaseIds
}
