% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-example.R
\name{getExampleAnalysisSpec}
\alias{getExampleAnalysisSpec}
\title{Get design spec of example result database}
\usage{
getExampleAnalysisSpec()
}
\value{
The json of the analysis spec for the example results
}
\description{
Get design spec of example result database
}
\details{
Finds the location within the package of the analysis spec json with example results for 1) CohortGenerator, 
2) Characterization, 3) PatientLevelPrediction, 4) CohortMethod, 5) SelfControlledCaseSeries and 6) CohortIncidence
}
