% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-example.R
\name{getExampleConnectionDetails}
\alias{getExampleConnectionDetails}
\title{A connection details to an example result database}
\usage{
getExampleConnectionDetails()
}
\value{
The connection details to an example result database
}
\description{
A connection details to an example result database
}
\details{
Finds the location within the package of an sqlite database with example results for 1) CohortGenerator, 
2) Characterization, 3) PatientLevelPrediction, 4) CohortMethod, 5) SelfControlledCaseSeries and 6) CohortIncidence
}
\concept{Example}
