% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-incidenceRates.R
\name{getIncidenceRateRanges}
\alias{getIncidenceRateRanges}
\title{Global ranges for IR values}
\usage{
getIncidenceRateRanges(dataSource, minPersonYears = 0)
}
\description{
Global ranges for IR values
}
