% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-compareCharacterization.R
\name{getResultsCohort}
\alias{getResultsCohort}
\title{Returns data from cohort table of Cohort Diagnostics results data model}
\usage{
getResultsCohort(dataSource, cohortIds = NULL)
}
\value{
Returns a data frame (tibble)
}
\description{
Returns data from cohort table of Cohort Diagnostics results data model

template DataSource

template CohortIds
}
