% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-cohort-overlap.R
\name{getResultsCohortOverlap}
\alias{getResultsCohortOverlap}
\title{Returns data for use in cohort_overlap}
\usage{
getResultsCohortOverlap(
  dataSource,
  targetCohortIds = NULL,
  comparatorCohortIds = NULL,
  databaseIds = NULL
)
}
\arguments{
\item{targetCohortIds}{A vector of cohort ids representing target cohorts}

\item{comparatorCohortIds}{A vector of cohort ids representing comparator cohorts}
}
\value{
Returns data for use in cohort_overlap
}
\description{
Returns data for use in cohort_overlap
}
