% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-cohort-overlap.R
\name{getResultsCohortRelationships}
\alias{getResultsCohortRelationships}
\title{Returns data from cohort_relationships table of Cohort Diagnostics results data model}
\usage{
getResultsCohortRelationships(
  dataSource,
  cohortIds = NULL,
  comparatorCohortIds = NULL,
  databaseIds = NULL,
  startDays = NULL,
  endDays = NULL
)
}
\arguments{
\item{startDays}{A vector of days in relation to cohort_start_date of target}

\item{endDays}{A vector of days in relation to cohort_end_date of target}
}
\value{
Returns a data frame (tibble) with results that conform to cohort_relationships
table in Cohort Diagnostics results data model.
}
\description{
Returns data from cohort_relationships table of Cohort Diagnostics results data model
}
