% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-sccsDataPulls.R
\name{getSccsExposuresOutcomes}
\alias{getSccsExposuresOutcomes}
\title{Get expsoure outcome pairs set for sccs module}
\usage{
getSccsExposuresOutcomes(
  connectionHandler,
  resultDatabaseSettings,
  includeControls = FALSE
)
}
\description{
Get expsoure outcome pairs set for sccs module
}
