% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-timeDistributions.R
\name{getTimeDistributionResult}
\alias{getTimeDistributionResult}
\title{Returns data from time_distribution table of Cohort Diagnostics results data model}
\usage{
getTimeDistributionResult(dataSource, cohortIds, databaseIds, databaseTable)
}
\value{
Returns a data frame (tibble).
}
\description{
Returns data from time_distribution table of Cohort Diagnostics results data model
}
