% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-inclusionRules.R
\name{inclusionRulesModule}
\alias{inclusionRulesModule}
\title{inclusion Rules Module}
\usage{
inclusionRulesModule(
  id,
  dataSource,
  databaseTable = dataSource$databaseTable,
  selectedCohort,
  targetCohortId,
  selectedDatabaseIds
)
}
\description{
inclusion Rules Module
}
