% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installModule.R
\name{installServer}
\alias{installServer}
\title{The server for installing ohdsi github packages}
\usage{
installServer(input, output, session, package)
}
\arguments{
\item{input}{Standard for shiny modules}

\item{output}{Standard for shiny modules}

\item{session}{Standard for shiny modules}

\item{package}{The name of the ohdsi github package}
}
\description{
The server for installing ohdsi github packages
}
\examples{
\dontrun{
webApi <- callModule(webApiServer, 'webApiMain')
cohortReactive <- callModule(OhdsiShinyModules::installServer, 'installPatientLevelPrediction',
                            package = 'PatientLevelPrediction')
}
}
