% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installModule.R
\name{installViewer}
\alias{installViewer}
\title{The user interface for installing ohdsi github packages}
\usage{
installViewer(id, label = "Package")
}
\arguments{
\item{id}{The module identifier}

\item{label}{The name of the ohdsi github package}
}
\description{
The user interface for installing ohdsi github packages
}
\examples{
\dontrun{
installViewer("installPatientLevelPrediction", "PatientLevelPrediction")
}
}
