% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-componentsCreateCustomColDefList.R
\name{makeButtonLabel}
\alias{makeButtonLabel}
\title{Make a label for an html button}
\usage{
makeButtonLabel(label)
}
\arguments{
\item{label}{The desired label for hte button}
}
\value{
html code to make a button label
}
\description{
Make a label for an html button
}
