% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelModule.R
\name{modelServer}
\alias{modelServer}
\title{The server for defining PatientLevelPrediction model settings}
\usage{
modelServer(input, output, session)
}
\arguments{
\item{input}{Standard for shiny modules}

\item{output}{Standard for shiny modules}

\item{session}{Standard for shiny modules}
}
\description{
The server for defining PatientLevelPrediction model settings
}
\examples{
\dontrun{
modelList <- callModule(modelServer, 'model1')
}
}
