% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelModule.R
\name{modelViewer}
\alias{modelViewer}
\title{The user interface for specifying the PatientLevelPrediction model}
\usage{
modelViewer(id, labelv = "Model")
}
\arguments{
\item{id}{The module identifier}

\item{labelv}{A string corresponding to the label for the form box}
}
\description{
The user interface for specifying the PatientLevelPrediction model
}
\examples{
\dontrun{
modelViewer('model1', labelv = "Model")
}
}
