% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-calibration.R
\name{patientLevelPredictionCalibrationViewer}
\alias{patientLevelPredictionCalibrationViewer}
\title{The module viewer for exploring prediction model calibration results}
\usage{
patientLevelPredictionCalibrationViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the prediction model calibration module
}
\description{
The module viewer for exploring prediction model calibration results
}
\details{
The user specifies the id for the module
}
