% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-designSummary.R
\name{patientLevelPredictionDesignSummaryServer}
\alias{patientLevelPredictionDesignSummaryServer}
\title{The module server for exploring prediction designs in the results}
\usage{
patientLevelPredictionDesignSummaryServer(
  id,
  connectionHandler,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
The server to the prediction design module
}
\description{
The module server for exploring prediction designs in the results
}
\details{
The user specifies the id for the module
}
