% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-designSummary.R
\name{patientLevelPredictionDesignSummaryViewer}
\alias{patientLevelPredictionDesignSummaryViewer}
\title{The module viewer for exploring prediction designs that have been run}
\usage{
patientLevelPredictionDesignSummaryViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the prediction design module
}
\description{
The module viewer for exploring prediction designs that have been run
}
\details{
The user specifies the id for the module
}
