% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-discrimination.R
\name{patientLevelPredictionDiscriminationViewer}
\alias{patientLevelPredictionDiscriminationViewer}
\title{The module viewer for exploring prediction model discrimination results}
\usage{
patientLevelPredictionDiscriminationViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the model discrimination results module
}
\description{
The module viewer for exploring prediction model discrimination results
}
\details{
The user specifies the id for the module
}
