% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-heatmap.R
\name{patientLevelPredictionHeatmapServer}
\alias{patientLevelPredictionHeatmapServer}
\title{The module server for exploring prediction model discrimination results}
\usage{
patientLevelPredictionHeatmapServer(
  id,
  performances,
  performanceRowIds,
  connectionHandler,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{performances}{a reactive val with the performance data.frame}

\item{performanceRowIds}{a vector of selected performance rows}

\item{connectionHandler}{the connection to the prediction result database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
The server to the model discrimination module
}
\description{
The module server for exploring prediction model discrimination results
}
\details{
The user specifies the id for the module
}
\seealso{
Other PatientLevelPrediction:
\code{\link[=patientLevelPredictionHeatmapViewer]{patientLevelPredictionHeatmapViewer()}},
\code{\link[=patientLevelPredictionHelperFile]{patientLevelPredictionHelperFile()}},
\code{\link[=patientLevelPredictionModelServer]{patientLevelPredictionModelServer()}},
\code{\link[=patientLevelPredictionModelViewer]{patientLevelPredictionModelViewer()}},
\code{\link[=patientLevelPredictionPlotServer]{patientLevelPredictionPlotServer()}},
\code{\link[=patientLevelPredictionPlotViewer]{patientLevelPredictionPlotViewer()}},
\code{\link[=patientLevelPredictionServer]{patientLevelPredictionServer()}},
\code{\link[=patientLevelPredictionViewer]{patientLevelPredictionViewer()}}
}
\concept{PatientLevelPrediction}
