% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-modelSummary.R
\name{patientLevelPredictionModelSummaryServer}
\alias{patientLevelPredictionModelSummaryServer}
\title{The module server for exploring prediction summary results}
\usage{
patientLevelPredictionModelSummaryServer(
  id,
  connectionHandler,
  resultDatabaseSettings,
  modelDesignId
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}

\item{modelDesignId}{a reactable id specifying the prediction model design identifier}
}
\value{
The server to the summary module
}
\description{
The module server for exploring prediction summary results
}
\details{
The user specifies the id for the module
}
