% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-netbenefit.R
\name{patientLevelPredictionNbServer}
\alias{patientLevelPredictionNbServer}
\title{The module server for exploring prediction net-benefit results}
\usage{
patientLevelPredictionNbServer(
  id,
  performanceId,
  connectionHandler,
  inputSingleView,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{performanceId}{the performance id in the database}

\item{connectionHandler}{the connection to the prediction result database}

\item{inputSingleView}{the current tab}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
The server to the net-benefit module
}
\description{
The module server for exploring prediction net-benefit results
}
\details{
The user specifies the id for the module
}
