% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-netbenefit.R
\name{patientLevelPredictionNbViewer}
\alias{patientLevelPredictionNbViewer}
\title{The module viewer for exploring prediction net-benefit results}
\usage{
patientLevelPredictionNbViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the net-benefit module
}
\description{
The module viewer for exploring prediction net-benefit results
}
\details{
The user specifies the id for the module
}
