% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-settings.R
\name{patientLevelPredictionSettingsServer}
\alias{patientLevelPredictionSettingsServer}
\title{The module server for exploring prediction settings}
\usage{
patientLevelPredictionSettingsServer(
  id,
  modelDesignId,
  developmentDatabaseId,
  performanceId,
  connectionHandler,
  inputSingleView,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{modelDesignId}{unique id for the model design}

\item{developmentDatabaseId}{unique id for the development database}

\item{performanceId}{unique id for the performance results}

\item{connectionHandler}{the connection to the prediction result database}

\item{inputSingleView}{the current tab}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
The server to the settings module
}
\description{
The module server for exploring prediction settings
}
\details{
The user specifies the id for the module
}
