% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-settings.R
\name{patientLevelPredictionSettingsViewer}
\alias{patientLevelPredictionSettingsViewer}
\title{The module viewer for exploring prediction settings}
\usage{
patientLevelPredictionSettingsViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the settings module
}
\description{
The module viewer for exploring prediction settings
}
\details{
The user specifies the id for the module
}
