% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-validation.R
\name{patientLevelPredictionValidationServer}
\alias{patientLevelPredictionValidationServer}
\title{The module server for exploring prediction validation results}
\usage{
patientLevelPredictionValidationServer(
  id,
  modelDesignId,
  developmentDatabaseId,
  performanceId,
  connectionHandler,
  inputSingleView,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{modelDesignId}{identifier for the model design}

\item{developmentDatabaseId}{identifier for the development database}

\item{performanceId}{identifier for the performance}

\item{connectionHandler}{the connection to the prediction result database}

\item{inputSingleView}{the current tab}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
The server to the validation module
}
\description{
The module server for exploring prediction validation results
}
\details{
The user specifies the id for the module
}
