% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-validation.R
\name{patientLevelPredictionValidationViewer}
\alias{patientLevelPredictionValidationViewer}
\title{The module viewer for exploring prediction validation results}
\usage{
patientLevelPredictionValidationViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the validation module
}
\description{
The module viewer for exploring prediction validation results
}
\details{
The user specifies the id for the module
}
