% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-main.R
\name{patientLevelPredictionViewer}
\alias{patientLevelPredictionViewer}
\title{The module viewer for exploring PatientLevelPrediction}
\usage{
patientLevelPredictionViewer(id = 1)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the PatientLevelPrediction viewer module
}
\description{
The module viewer for exploring PatientLevelPrediction
}
\details{
The user specifies the id for the module
}
\seealso{
Other PatientLevelPrediction:
\code{\link[=patientLevelPredictionHeatmapServer]{patientLevelPredictionHeatmapServer()}},
\code{\link[=patientLevelPredictionHeatmapViewer]{patientLevelPredictionHeatmapViewer()}},
\code{\link[=patientLevelPredictionHelperFile]{patientLevelPredictionHelperFile()}},
\code{\link[=patientLevelPredictionModelServer]{patientLevelPredictionModelServer()}},
\code{\link[=patientLevelPredictionModelViewer]{patientLevelPredictionModelViewer()}},
\code{\link[=patientLevelPredictionPlotServer]{patientLevelPredictionPlotServer()}},
\code{\link[=patientLevelPredictionPlotViewer]{patientLevelPredictionPlotViewer()}},
\code{\link[=patientLevelPredictionServer]{patientLevelPredictionServer()}}
}
\concept{PatientLevelPrediction}
