% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phevaluator-main.R
\name{phevaluatorHelperFile}
\alias{phevaluatorHelperFile}
\title{The location of the phevaluator module helper file}
\usage{
phevaluatorHelperFile()
}
\value{
String location of the phevaluator helper file
}
\description{
The location of the phevaluator module helper file
}
\details{
Returns the location of the cohort-generator helper file
}
\seealso{
Other PheValuator:
\code{\link[=phevaluatorServer]{phevaluatorServer()}},
\code{\link[=phevaluatorViewer]{phevaluatorViewer()}}
}
\concept{PheValuator}
