% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationModule.R
\name{populationServer}
\alias{populationServer}
\title{The server for creating a study population (using PatientLevelPrediction)}
\usage{
populationServer(input, output, session)
}
\arguments{
\item{input}{Standard for shiny modules}

\item{output}{Standard for shiny modules}

\item{session}{Standard for shiny modules}
}
\description{
The server for creating a study population (using PatientLevelPrediction)
}
\examples{
\dontrun{
popList <- callModule(populationServer, 'population1')
}
}
