% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populationModule.R
\name{populationViewer}
\alias{populationViewer}
\title{The user interface for defining study population settings (using PatientLevelPrediction)}
\usage{
populationViewer(id, labelv = "Population")
}
\arguments{
\item{id}{The module identifier}

\item{labelv}{A string used as the label to the box with the population settings form}
}
\description{
The user interface for defining study population settings (using PatientLevelPrediction)
}
\examples{
\dontrun{
populationViewer('population1', labelv = "Population")
}
}
