% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-calibration.R
\name{predictionCalibrationServer}
\alias{predictionCalibrationServer}
\title{The module server for exploring prediction validation results}
\usage{
predictionCalibrationServer(
  id,
  performanceId,
  connectionHandler,
  inputSingleView,
  mySchema,
  myTableAppend
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{performanceId}{the performance id in the database}

\item{connectionHandler}{the connection to the prediction result database}

\item{inputSingleView}{the current tab}

\item{mySchema}{the database schema for the model results}

\item{myTableAppend}{a string that appends the tables in the result schema}
}
\value{
The server to the prediction calibration module
}
\description{
The module server for exploring prediction validation results
}
\details{
The user specifies the id for the module
}
