% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-calibration.R
\name{predictionCalibrationViewer}
\alias{predictionCalibrationViewer}
\title{The module viewer for exploring prediction model calibration results}
\usage{
predictionCalibrationViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the prediction model calibration module
}
\description{
The module viewer for exploring prediction model calibration results
}
\details{
The user specifies the id for the module
}
