% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-covariateSummary.R
\name{predictionCovariateSummaryViewer}
\alias{predictionCovariateSummaryViewer}
\title{The module viewer for exploring prediction covariate summary results}
\usage{
predictionCovariateSummaryViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the covariate summary module
}
\description{
The module viewer for exploring prediction covariate summary results
}
\details{
The user specifies the id for the module
}
