% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-cutoff.R
\name{predictionCutoffViewer}
\alias{predictionCutoffViewer}
\title{The module viewer for exploring prediction cut-off results}
\usage{
predictionCutoffViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the prediction cut-off module
}
\description{
The module viewer for exploring prediction cut-off results
}
\details{
The user specifies the id for the module
}
