% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-diagnostics.R
\name{predictionDiagnosticsServer}
\alias{predictionDiagnosticsServer}
\title{The module server for exploring prediction diagnostic results}
\usage{
predictionDiagnosticsServer(
  id,
  modelDesignId,
  schema,
  connectionHandler,
  plpTablePrefix,
  databaseTablePrefix
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{modelDesignId}{the unique id for the model design}

\item{schema}{the database schema for the model results}

\item{connectionHandler}{the connection to the prediction result database}

\item{plpTablePrefix}{a string that appends the tables in the result schema}

\item{databaseTablePrefix}{a string that appends the database_meta_data table}
}
\value{
The server to the predcition diagnostic module
}
\description{
The module server for exploring prediction diagnostic results
}
\details{
The user specifies the id for the module
}
