% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-diagnostics.R
\name{predictionDiagnosticsViewer}
\alias{predictionDiagnosticsViewer}
\title{The module viewer for exploring prediction diagnostic results}
\usage{
predictionDiagnosticsViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the prediction diagnostic module
}
\description{
The module viewer for exploring prediction diagnostic results
}
\details{
The user specifies the id for the module
}
