% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-main.R
\name{predictionHelperFile}
\alias{predictionHelperFile}
\title{The location of the prediction module helper file}
\usage{
predictionHelperFile()
}
\value{
string location of the prediction helper file
}
\description{
The location of the prediction module helper file
}
\details{
Returns the location of the prediction helper file
}
