% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-modelSummary.R
\name{predictionModelSummaryViewer}
\alias{predictionModelSummaryViewer}
\title{The module viewer for exploring prediction summary results}
\usage{
predictionModelSummaryViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the summary module
}
\description{
The module viewer for exploring prediction summary results
}
\details{
The user specifies the id for the module
}
