% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-netbenefit.R
\name{predictionNbServer}
\alias{predictionNbServer}
\title{The module server for exploring prediction net-benefit results}
\usage{
predictionNbServer(
  id,
  performanceId,
  connectionHandler,
  inputSingleView,
  schema,
  plpTablePrefix
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{performanceId}{the performance id in the database}

\item{connectionHandler}{the connection to the prediction result database}

\item{inputSingleView}{the current tab}

\item{schema}{the database schema for the model results}

\item{plpTablePrefix}{a string that appends the tables in the result schema}
}
\value{
The server to the net-benefit module
}
\description{
The module server for exploring prediction net-benefit results
}
\details{
The user specifies the id for the module
}
