% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-netbenefit.R
\name{predictionNbViewer}
\alias{predictionNbViewer}
\title{The module viewer for exploring prediction net-benefit results}
\usage{
predictionNbViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the net-benefit module
}
\description{
The module viewer for exploring prediction net-benefit results
}
\details{
The user specifies the id for the module
}
