% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-main.R
\name{predictionServer}
\alias{predictionServer}
\title{The module server for exploring PatientLevelPrediction}
\usage{
predictionServer(id, resultDatabaseSettings = list(myPort = 1))
}
\arguments{
\item{id}{the unique reference id for the module}

\item{resultDatabaseSettings}{a list containing the prediction result schema and connection details}
}
\value{
The server for the PatientLevelPrediction module
}
\description{
The module server for exploring PatientLevelPrediction
}
\details{
The user specifies the id for the module
}
