% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-settings.R
\name{predictionSetingsServer}
\alias{predictionSetingsServer}
\title{The module server for exploring prediction settings}
\usage{
predictionSetingsServer(
  id,
  modelDesignId,
  developmentDatabaseId,
  performanceId,
  mySchema,
  con,
  inputSingleView,
  myTableAppend,
  targetDialect
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{modelDesignId}{unique id for the model design}

\item{developmentDatabaseId}{unique id for the development database}

\item{performanceId}{unique id for the performance results}

\item{mySchema}{the database schema for the model results}

\item{con}{the connection to the prediction result database}

\item{inputSingleView}{the current tab}

\item{myTableAppend}{a string that appends the tables in the result schema}

\item{targetDialect}{the database management system for the model results}
}
\value{
The server to the settings module
}
\description{
The module server for exploring prediction settings
}
\details{
The user specifies the id for the module
}
