% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-settings.R
\name{predictionSettingsViewer}
\alias{predictionSettingsViewer}
\title{The module viewer for exploring prediction settings}
\usage{
predictionSettingsViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the settings module
}
\description{
The module viewer for exploring prediction settings
}
\details{
The user specifies the id for the module
}
