% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-main.R
\name{predictionViewer}
\alias{predictionViewer}
\title{The module viewer for exploring PatientLevelPrediction}
\usage{
predictionViewer(id = 1)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the PatientLevelPrediction viewer module
}
\description{
The module viewer for exploring PatientLevelPrediction
}
\details{
The user specifies the id for the module
}
