% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-main.R
\name{reportHelperFile}
\alias{reportHelperFile}
\title{The location of the report module helper file}
\usage{
reportHelperFile()
}
\value{
string location of the report helper file
}
\description{
The location of the report module helper file
}
\details{
Returns the location of the report helper file
}
\seealso{
Other {Report}: 
\code{\link{reportServer}()},
\code{\link{reportViewer}()}
}
\concept{{Report}}
