% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-data-viewer.R
\name{resultTableViewer}
\alias{resultTableViewer}
\title{Result Table Viewer}
\usage{
resultTableViewer(id = "result-table", boxTitle = "Table")
}
\arguments{
\item{id}{string}

\item{boxTitle}{the title added to the box}
}
\value{
shiny module UI
}
\description{
Result Table Viewer
}
\seealso{
Other Utils:
\code{\link[=actionButtonStyleInfo]{actionButtonStyleInfo()}},
\code{\link[=actionButtonStyleWarning]{actionButtonStyleWarning()}},
\code{\link[=createActionButton]{createActionButton()}},
\code{\link[=datasourcesHelperFile]{datasourcesHelperFile()}},
\code{\link[=datasourcesServer]{datasourcesServer()}},
\code{\link[=datasourcesViewer]{datasourcesViewer()}},
\code{\link[=getLogoImage]{getLogoImage()}}
}
\concept{Utils}
