% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heplers-componentsCreateCustomColDefList.R
\name{saveColDefsAsJSON}
\alias{saveColDefsAsJSON}
\title{Saving a custom colDef list to a JSON file}
\usage{
saveColDefsAsJSON(colDefs, filename, path = "")
}
\arguments{
\item{colDefs}{A named list of reactable::colDef objects, generated by createCustomColDefList()}

\item{filename}{The desired name of the output JSON file}

\item{path}{The path to where the JSON should be saved}
}
\value{
Saves a JSON file
}
\description{
Saving a custom colDef list to a JSON file
}
