% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccs-diagnosticsSummary.R
\name{sccsDiagnosticsSummaryServer}
\alias{sccsDiagnosticsSummaryServer}
\title{The module server for rendering the SCCS diagnostics summary}
\usage{
sccsDiagnosticsSummaryServer(id, connectionHandler, resultDatabaseSettings)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultDatabaseSettings}{the resultDatabaseSettings with the schemas, prefix and table names}
}
\value{
the SCCS diagnostics summary results
}
\description{
The module server for rendering the SCCS diagnostics summary
}
