% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccs-diagnosticsSummary.R
\name{sccsDiagnosticsSummaryViewer}
\alias{sccsDiagnosticsSummaryViewer}
\title{The module viewer for rendering the SCCS diagnostics results}
\usage{
sccsDiagnosticsSummaryViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation diagnostics viewer
}
\description{
The module viewer for rendering the SCCS diagnostics results
}
