% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sccs-main.R
\name{sccsHelperFile}
\alias{sccsHelperFile}
\title{The location of the description module helper file}
\usage{
sccsHelperFile()
}
\value{
string location of the description helper file
}
\description{
The location of the description module helper file
}
\details{
Returns the location of the description helper file
}
