% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-pair.R
\name{selfControlledCohortDiagnosticsColDef}
\alias{selfControlledCohortDiagnosticsColDef}
\title{The column definitions for the per database study diagnostics table}
\usage{
selfControlledCohortDiagnosticsColDef()
}
\value{
A named list of reactable::colDef
}
\description{
The column definitions for the per database study diagnostics table
}
\details{
The diagnostics come from the scc_diagnostics_summary table (MDRR, EASE,
pre exposure tests etc) and the summaryValue column reflects whether the
pair passed the diagnostics in each database
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsViewer}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
