% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-main.R
\name{selfControlledCohortHelperFile}
\alias{selfControlledCohortHelperFile}
\title{The location of the self controlled cohort module helper file}
\usage{
selfControlledCohortHelperFile()
}
\value{
string location of the self controlled cohort helper file
}
\description{
The location of the self controlled cohort module helper file
}
\details{
Returns the location of the self controlled cohort helper file
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortMetaSql}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsSql}()},
\code{\link{selfControlledCohortSignalsViewer}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
