% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-pair.R
\name{selfControlledCohortMetaDiagnosticsColDef}
\alias{selfControlledCohortMetaDiagnosticsColDef}
\title{The column definitions for the pair level meta analysis diagnostics}
\usage{
selfControlledCohortMetaDiagnosticsColDef()
}
\value{
A named list of reactable::colDef
}
\description{
The column definitions for the pair level meta analysis diagnostics
}
\details{
Shows the evidence synthesis diagnostic values and their PASS/FAIL status so
the reason a meta analytic result is blinded (failed a diagnostic) is
visible
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortMetaSql}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsSql}()},
\code{\link{selfControlledCohortSignalsViewer}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
