% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-pair.R
\name{selfControlledCohortPairServer}
\alias{selfControlledCohortPairServer}
\title{The module server for exploring a single exposure-outcome pair}
\usage{
selfControlledCohortPairServer(
  id,
  connectionHandler,
  resultDatabaseSettings,
  selectedPair
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list with the result schema and table
prefixes}

\item{selectedPair}{a reactiveVal holding the selected pair (a one row
data.frame with targetId, targetName, outcomeId and outcomeName)}
}
\value{
The server for the pair explorer module
}
\description{
The module server for exploring a single exposure-outcome pair
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaDiagnosticsColDef}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortMetaSql}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsSql}()},
\code{\link{selfControlledCohortSignalsViewer}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
