% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-pair.R
\name{selfControlledCohortPairViewer}
\alias{selfControlledCohortPairViewer}
\title{The module viewer for exploring a single exposure-outcome pair}
\usage{
selfControlledCohortPairViewer(id = "pair")
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the pair explorer module
}
\description{
The module viewer for exploring a single exposure-outcome pair
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaDiagnosticsColDef}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortMetaSql}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsSql}()},
\code{\link{selfControlledCohortSignalsViewer}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
