% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-signals.R
\name{selfControlledCohortSignalsViewer}
\alias{selfControlledCohortSignalsViewer}
\title{The module viewer for the self controlled cohort signal discovery grid}
\usage{
selfControlledCohortSignalsViewer(id = "signals")
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the signal discovery grid
}
\description{
The module viewer for the self controlled cohort signal discovery grid
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortMetaSql}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsSql}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
