% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment-patterns-main.R
\name{treatmentPatternsHelperFile}
\alias{treatmentPatternsHelperFile}
\title{The location of the treatmentPatterns module helper file}
\usage{
treatmentPatternsHelperFile()
}
\value{
string location of the treatment-patterns helper file
}
\description{
The location of the treatmentPatterns module helper file
}
\details{
Returns the location of the treatment-patterns helper file
}
\seealso{
Other Treatment Patterns: 
\code{\link{treatmentPatternsServer}()},
\code{\link{treatmentPatternsViewer}()}
}
\concept{Treatment Patterns}
