% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment-patterns-main.R
\name{treatmentPatternsServer}
\alias{treatmentPatternsServer}
\title{The module server for exploring treatment patterns studies}
\usage{
treatmentPatternsServer(id, connectionHandler, resultDatabaseSettings)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list containing the characterization result schema, dbms, tablePrefix, databaseTable and cgTablePrefix}
}
\value{
The server for the treatment-patterns module
}
\description{
The module server for exploring treatment patterns studies
}
\details{
The user specifies the id for the module
}
\seealso{
Other Treatment Patterns: 
\code{\link{treatmentPatternsHelperFile}()},
\code{\link{treatmentPatternsViewer}()}
}
\concept{Treatment Patterns}
