% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treatment-patterns-main.R
\name{treatmentPatternsViewer}
\alias{treatmentPatternsViewer}
\title{The module viewer for exploring treatment patterns studies}
\usage{
treatmentPatternsViewer(id = 1)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the treatment-patterns viewer module
}
\description{
The module viewer for exploring treatment patterns studies
}
\details{
The user specifies the id for the module
}
\seealso{
Other Treatment Patterns: 
\code{\link{treatmentPatternsHelperFile}()},
\code{\link{treatmentPatternsServer}()}
}
\concept{Treatment Patterns}
