% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webApiModule.R
\name{webApiServer}
\alias{webApiServer}
\title{The server for connecting to an ATLAS webapi}
\usage{
webApiServer(input, output, session)
}
\arguments{
\item{input}{Standard for shiny modules}

\item{output}{Standard for shiny modules}

\item{session}{Standard for shiny modules}
}
\description{
The server for connecting to an ATLAS webapi
}
\examples{
\dontrun{
webApi <- callModule(webApiServer, 'webApi1')
}
}
