% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webApiModule.R
\name{webApiViewer}
\alias{webApiViewer}
\title{The user interface for connecting to an ATLAS webapi}
\usage{
webApiViewer(id)
}
\arguments{
\item{id}{The module identifier}
}
\description{
The user interface for connecting to an ATLAS webapi
}
\examples{
\dontrun{
webApiViewer('webApi1')
}
}
