% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-actions.R
\name{actionButtonStyleInfo}
\alias{actionButtonStyleInfo}
\title{Action Button Style (Info)}
\usage{
actionButtonStyleInfo()
}
\value{
A CSS style string.
}
\description{
Shared styling for neutral/info action pills used in result tables.
}
\seealso{
Other Utils:
\code{\link[=actionButtonStyleWarning]{actionButtonStyleWarning()}},
\code{\link[=createActionButton]{createActionButton()}},
\code{\link[=datasourcesHelperFile]{datasourcesHelperFile()}},
\code{\link[=datasourcesServer]{datasourcesServer()}},
\code{\link[=datasourcesViewer]{datasourcesViewer()}},
\code{\link[=getLogoImage]{getLogoImage()}},
\code{\link[=resultTableViewer]{resultTableViewer()}}
}
\concept{Utils}
