% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-dechallengeRechallenge.R
\name{characterizationDechallengeRechallengeServer}
\alias{characterizationDechallengeRechallengeServer}
\title{The module server for exploring Dechallenge Rechallenge results}
\usage{
characterizationDechallengeRechallengeServer(
  id,
  connectionHandler,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{resultDatabaseSettings}{a list containing the characterization result schema, dbms, tablePrefix, databaseTable and cgTablePrefix}
}
\value{
The server to the Dechallenge Rechallenge module
}
\description{
The module server for exploring Dechallenge Rechallenge results
}
\details{
The user specifies the id for the module
}
