% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-incidence.R
\name{characterizationIncidenceServer}
\alias{characterizationIncidenceServer}
\title{The module server for exploring incidence results}
\usage{
characterizationIncidenceServer(
  id,
  connectionHandler,
  mainPanelTab,
  schema,
  incidenceTablePrefix,
  databaseTable = "DATABASE_META_DATA"
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{mainPanelTab}{the current tab}

\item{schema}{the database schema for the model results}

\item{incidenceTablePrefix}{a string that appends the incidence table in the result schema}

\item{databaseTable}{name of the database table}
}
\value{
The server to the prediction incidence module
}
\description{
The module server for exploring incidence results
}
\details{
The user specifies the id for the module
}
