% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-incidence.R
\name{characterizationIncidenceServer}
\alias{characterizationIncidenceServer}
\title{The module server for exploring incidence results}
\usage{
characterizationIncidenceServer(
  id,
  connectionHandler,
  resultDatabaseSettings,
  options,
  parents,
  parentIndex,
  outcomes,
  targetIds
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{resultDatabaseSettings}{a list containing the characterization result schema, dbms, tablePrefix, databaseTable and cgTablePrefix}
}
\value{
The server to the prediction incidence module
}
\description{
The module server for exploring incidence results
}
\details{
The user specifies the id for the module
}
