% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/characterization-main.R
\name{characterizationViewer}
\alias{characterizationViewer}
\title{The module viewer for exploring characterization studies}
\usage{
characterizationViewer(id = 1)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the characterization viewer module
}
\description{
The module viewer for exploring characterization studies
}
\details{
The user specifies the id for the module
}
\seealso{
Other Characterization: 
\code{\link{characterizationHelperFile}()},
\code{\link{characterizationIncidenceServer}()},
\code{\link{characterizationIncidenceViewer}()},
\code{\link{characterizationServer}()}
}
\concept{Characterization}
