% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-counts.R
\name{cohortCountsModule}
\alias{cohortCountsModule}
\title{Shiny module for cohort counts}
\usage{
cohortCountsModule(
  id,
  dataSource,
  cohortTable = dataSource$cohortTable,
  databaseTable = dataSource$dbTable,
  selectedCohorts,
  selectedDatabaseIds,
  cohortIds
)
}
\arguments{
\item{id}{namespace id}

\item{dataSource}{Backend Data source (DatabaseConnection)}

\item{cohortTable}{data.frame of all cohorts}

\item{databaseTable}{data.frame of all databases}

\item{selectedCohorts}{shiny::reactive - should return cohorts selected or NULL}

\item{selectedDatabaseIds}{shiny::reactive - should return cohorts selected or NULL}

\item{cohortIds}{shiny::reactive - should return cohorts selected integers or NULL}
}
\description{
Shiny module for cohort counts. Displays reactable table of cohort counts
}
\seealso{
Other {CohortDiagnostics}: 
\code{\link{cohortCountsView}()},
\code{\link{cohortDefinitionsModule}()},
\code{\link{cohortDefinitionsView}()},
\code{\link{cohortDiagCharacterizationView}()},
\code{\link{cohortDiagnosticsHelperFile}()},
\code{\link{cohortDiagnosticsServer}()},
\code{\link{cohortDiagnosticsView}()},
\code{\link{cohortOverlapView}()},
\code{\link{compareCohortCharacterizationView}()},
\code{\link{conceptsInDataSourceView}()},
\code{\link{createCdDatabaseDataSource}()},
\code{\link{databaseInformationView}()},
\code{\link{getCirceRenderedExpression}()},
\code{\link{getEnabledCdReports}()},
\code{\link{incidenceRatesView}()},
\code{\link{inclusionRulesView}()},
\code{\link{indexEventBreakdownView}()},
\code{\link{orpahanConceptsView}()},
\code{\link{timeDistributionsView}()},
\code{\link{visitContextView}()}
}
\concept{{CohortDiagnostics}}
