% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-counts.R
\name{cohortCountsView}
\alias{cohortCountsView}
\title{Cohort Counts View}
\usage{
cohortCountsView(id)
}
\arguments{
\item{id}{Namespace id}
}
\description{
Shiny view for cohort counts module
}
