% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-counts.R
\name{cohortCountsView}
\alias{cohortCountsView}
\title{Cohort Counts View}
\usage{
cohortCountsView(id)
}
\description{
Shiny view for cohort counts module
inputId cohortCountsTableColumnFilter              Column filters
outputId cohortCountsTable                         Reactable output of cohort counts for specified databases
outputId inclusionRuleStats    Reactable output of inclusion rules
}
