% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-counts.R
\name{cohortCountsView}
\alias{cohortCountsView}
\title{Cohort Counts View}
\usage{
cohortCountsView(id)
}
\arguments{
\item{id}{Namespace id}
}
\description{
Shiny view for cohort counts module
}
\seealso{
Other CohortDiagnostics: 
\code{\link{cohortCountsModule}()},
\code{\link{cohortDefinitionsModule}()},
\code{\link{cohortDefinitionsView}()},
\code{\link{cohortDiagCharacterizationView}()},
\code{\link{cohortDiagnosticsHelperFile}()},
\code{\link{cohortDiagnosticsServer}()},
\code{\link{cohortDiagnosticsView}()},
\code{\link{cohortOverlapView}()},
\code{\link{compareCohortCharacterizationView}()},
\code{\link{conceptsInDataSourceView}()},
\code{\link{createCdDatabaseDataSource}()},
\code{\link{databaseInformationView}()},
\code{\link{getCirceRenderedExpression}()},
\code{\link{getEnabledCdReports}()},
\code{\link{incidenceRatesView}()},
\code{\link{inclusionRulesView}()},
\code{\link{indexEventBreakdownView}()},
\code{\link{orpahanConceptsView}()},
\code{\link{timeDistributionsView}()},
\code{\link{visitContextView}()}
}
\concept{CohortDiagnostics}
