% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-definition.R
\name{cohortDefinitionsModule}
\alias{cohortDefinitionsModule}
\title{Cohort Definition module}
\usage{
cohortDefinitionsModule(
  id,
  dataSource,
  cohortDefinitions,
  cohortTable = dataSource$cohortTable,
  cohortCount = dataSource$cohortCountTable,
  databaseTable = dataSource$databaseTable
)
}
\arguments{
\item{id}{Namespace id}

\item{dataSource}{DatabaseConnection}

\item{cohortDefinitions}{reactive of cohort definitions to display}

\item{databaseTable}{data.frame of databasese}
}
\description{
cohort defintion conceptsets, json etc
}
