% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-definition.R
\name{cohortDefinitionsView}
\alias{cohortDefinitionsView}
\title{Cohort Definitions View}
\usage{
cohortDefinitionsView(id, parentId = "DiagnosticsExplorer")
}
\arguments{
\item{id}{Namespace id for module}
}
\description{
Outputs cohort definitions
}
