% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main-ui.R
\name{cohortDiagnosticsHelperFile}
\alias{cohortDiagnosticsHelperFile}
\title{The location of the description module helper file}
\usage{
cohortDiagnosticsHelperFile()
}
\value{
string location of the description helper file
}
\description{
The location of the description module helper file
}
\details{
Returns the location of the description helper file
}
\seealso{
Other {CohortDiagnostics}: 
\code{\link{cohortDiagnosticsServer}()},
\code{\link{orpahanConceptsView}()}
}
\concept{{CohortDiagnostics}}
