% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main-ui.R
\name{cohortDiagnosticsHelperFile}
\alias{cohortDiagnosticsHelperFile}
\title{The location of the description module helper file}
\usage{
cohortDiagnosticsHelperFile()
}
\value{
string location of the description helper file
}
\description{
The location of the description module helper file
}
\details{
Returns the location of the description helper file
}
\seealso{
Other {CohortDiagnostics}: 
\code{\link{cohortCountsModule}()},
\code{\link{cohortCountsView}()},
\code{\link{cohortDefinitionsModule}()},
\code{\link{cohortDefinitionsView}()},
\code{\link{cohortDiagCharacterizationView}()},
\code{\link{cohortDiagnosticsServer}()},
\code{\link{cohortDiagnosticsView}()},
\code{\link{cohortOverlapView}()},
\code{\link{compareCohortCharacterizationView}()},
\code{\link{conceptsInDataSourceView}()},
\code{\link{createCdDatabaseDataSource}()},
\code{\link{databaseInformationView}()},
\code{\link{getCirceRenderedExpression}()},
\code{\link{getEnabledCdReports}()},
\code{\link{incidenceRatesView}()},
\code{\link{inclusionRulesView}()},
\code{\link{indexEventBreakdownView}()},
\code{\link{orpahanConceptsView}()},
\code{\link{timeDistributionsView}()},
\code{\link{visitContextView}()}
}
\concept{{CohortDiagnostics}}
