% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main.R
\name{cohortDiagnosticsServer}
\alias{cohortDiagnosticsServer}
\title{Cohort Diagnostics Explorer main module}
\usage{
cohortDiagnosticsServer(
  id,
  connectionHandler,
  resultDatabaseSettings,
  dataSource = NULL
)
}
\arguments{
\item{id}{module Id}

\item{connectionHandler}{ResultModelManager ConnectionHander instance}

\item{resultDatabaseSettings}{results database settings}

\item{dataSource}{dataSource optionally created with createCdDatabaseDataSource}
}
\description{
Cohort Diagnostics Explorer main module
}
\seealso{
Other {CohortDiagnostics}: 
\code{\link{cohortCountsModule}()},
\code{\link{cohortCountsView}()},
\code{\link{cohortDefinitionsModule}()},
\code{\link{cohortDefinitionsView}()},
\code{\link{cohortDiagCharacterizationView}()},
\code{\link{cohortDiagnosticsHelperFile}()},
\code{\link{cohortDiagnosticsView}()},
\code{\link{cohortOverlapView}()},
\code{\link{compareCohortCharacterizationView}()},
\code{\link{conceptsInDataSourceView}()},
\code{\link{createCdDatabaseDataSource}()},
\code{\link{databaseInformationView}()},
\code{\link{getCirceRenderedExpression}()},
\code{\link{getEnabledCdReports}()},
\code{\link{incidenceRatesView}()},
\code{\link{inclusionRulesView}()},
\code{\link{indexEventBreakdownView}()},
\code{\link{orpahanConceptsView}()},
\code{\link{timeDistributionsView}()},
\code{\link{visitContextView}()}
}
\concept{{CohortDiagnostics}}
