% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main.R
\name{cohortDiagnosticsServer}
\alias{cohortDiagnosticsServer}
\title{Cohort Diagnostics Explorer main module}
\usage{
cohortDiagnosticsServer(
  id,
  connectionHandler,
  resultDatabaseSettings,
  dataSource = NULL
)
}
\arguments{
\item{id}{module Id}

\item{connectionHandler}{ResultModelManager ConnectionHander instance}

\item{resultDatabaseSettings}{results database settings}

\item{dataSource}{dataSource optionally created with createCdDatabaseDataSource}
}
\description{
Cohort Diagnostics Explorer main module
}
\seealso{
Other {CohortDiagnostics}: 
\code{\link{cohortDiagnosticsHelperFile}()},
\code{\link{orpahanConceptsView}()}
}
\concept{{CohortDiagnostics}}
