% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main.R
\name{cohortDiagnosticsSever}
\alias{cohortDiagnosticsSever}
\title{Cohort Diagnostics Explorer main module}
\usage{
cohortDiagnosticsSever(
  id,
  connectionHandler,
  resultDatabaseSettings,
  dataSource = NULL
)
}
\arguments{
\item{id}{module Id}

\item{connectionHandler}{ResultModelManager ConnectionHander instance}

\item{resultDatabaseSettings}{results database settings}

\item{dataSource}{dataSource optionally created with createCdDatabaseDataSource}
}
\description{
Cohort Diagnostics Explorer main module
}
