% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main.R
\name{cohortDiagnosticsSever}
\alias{cohortDiagnosticsSever}
\title{Cohort Diagnostics Explorer main module}
\usage{
cohortDiagnosticsSever(
  id = "DiagnosticsExplorer",
  connectionHandler,
  resultsDatabaseSettings,
  vocabularyDatabaseSchema = resultsDatabaseSettings$resultsDatabaseSchema
)
}
\arguments{
\item{connectionHandler}{ResultModelManager ConnectionHander instance}

\item{resultDatabaseSettings}{results database settings}
}
\description{
Cohort Diagnostics Explorer main module
}
