% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main-ui.R
\name{cohortDiagnosticsUi}
\alias{cohortDiagnosticsUi}
\title{Cohort Diagnostics UI}
\usage{
cohortDiagnosticsUi(id = "DiagnosticsExplorer", enabledTabs)
}
\arguments{
\item{id}{Namespace id "DiagnosticsExplorer"}

\item{enabledTabs}{enabled reports}
}
\description{
Cohort Diagnostics UI
}
