% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main-ui.R
\name{cohortDiagnosticsView}
\alias{cohortDiagnosticsView}
\title{View for cohort diagnostics module}
\usage{
cohortDiagnosticsView(id = "DiagnosticsExplorer")
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort diagnostics viewer module
}
\description{
View for cohort diagnostics module
}
\details{
The user specifies the id for the module
}
\seealso{
Other {CohortDiagnostics}: 
\code{\link{cohortDiagnosticsHelperFile}()},
\code{\link{cohortDiagnosticsServer}()},
\code{\link{orpahanConceptsView}()}
}
\concept{{CohortDiagnostics}}
