% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortgenerator-main.R
\name{cohortGeneratorHelperFile}
\alias{cohortGeneratorHelperFile}
\title{The location of the cohort-generator module helper file}
\usage{
cohortGeneratorHelperFile()
}
\value{
string location of the cohort-generator helper file
}
\description{
The location of the cohort-generator module helper file
}
\details{
Returns the location of the cohort-generator helper file
}
