% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohortgenerator-main.R
\name{cohortGeneratorServer}
\alias{cohortGeneratorServer}
\title{The module server for the main cohort generator module}
\usage{
cohortGeneratorServer(id, resultDatabaseSettings, resultsSchema = "poc")
}
\arguments{
\item{id}{the unique reference id for the module}

\item{resultDatabaseSettings}{a named list containing the cohort generator results database connection details}

\item{resultsSchema}{the schema with the cohort generator results}
}
\value{
the cohort generator results viewer main module server
}
\description{
The module server for the main cohort generator module
}
