% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-generator-main.R
\name{cohortGeneratorViewer}
\alias{cohortGeneratorViewer}
\title{The viewer of the main cohort generator module}
\usage{
cohortGeneratorViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort generator results viewer
}
\description{
The viewer of the main cohort generator module
}
\seealso{
Other CohortGenerator: 
\code{\link{cohortGeneratorHelperFile}()},
\code{\link{cohortGeneratorServer}()}
}
\concept{CohortGenerator}
