% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-attrition.R
\name{cohortMethodAttritionViewer}
\alias{cohortMethodAttritionViewer}
\title{The module viewer for rendering the PLE attrition results}
\usage{
cohortMethodAttritionViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method attrition
}
\description{
The module viewer for rendering the PLE attrition results
}
