% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-covariateBalance.R
\name{cohortMethodCovariateBalanceServer}
\alias{cohortMethodCovariateBalanceServer}
\title{The module server for rendering the covariate balance plot}
\usage{
cohortMethodCovariateBalanceServer(
  id,
  selectedRow,
  inputParams,
  connectionHandler,
  resultsSchema,
  tablePrefix,
  metaAnalysisDbIds = NULL
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultsSchema}{the schema with the PLE results}

\item{tablePrefix}{tablePrefix}

\item{metaAnalysisDbIds}{metaAnalysisDbIds}
}
\value{
the PLE covariate balance content server
}
\description{
The module server for rendering the covariate balance plot
}
