% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-covariateBalance.R
\name{cohortMethodCovariateBalanceServer}
\alias{cohortMethodCovariateBalanceServer}
\title{The module server for rendering the covariate balance plot}
\usage{
cohortMethodCovariateBalanceServer(
  id,
  selectedRow,
  connectionHandler,
  resultDatabaseSettings,
  metaAnalysisDbIds = NULL
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}

\item{metaAnalysisDbIds}{metaAnalysisDbIds}
}
\value{
the PLE covariate balance content server
}
\description{
The module server for rendering the covariate balance plot
}
\seealso{
Other Estimation: 
\code{\link{cohortMethodAttritionServer}()},
\code{\link{cohortMethodAttritionViewer}()},
\code{\link{cohortMethodCovariateBalanceViewer}()},
\code{\link{cohortMethodKaplanMeierServer}()},
\code{\link{cohortMethodKaplanMeierViewer}()},
\code{\link{cohortMethodPopulationCharacteristicsServer}()},
\code{\link{cohortMethodPopulationCharacteristicsViewer}()},
\code{\link{cohortMethodPowerServer}()},
\code{\link{cohortMethodPowerViewer}()},
\code{\link{cohortMethodPropensityModelServer}()},
\code{\link{cohortMethodPropensityModelViewer}()},
\code{\link{cohortMethodPropensityScoreDistServer}()},
\code{\link{cohortMethodPropensityScoreDistViewer}()},
\code{\link{cohortMethodSystematicErrorServer}()},
\code{\link{cohortMethodSystematicErrorViewer}()},
\code{\link{estimationHelperFile}()},
\code{\link{estimationServer}()},
\code{\link{estimationViewer}()}
}
\concept{Estimation}
