% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-covariateBalance.R
\name{cohortMethodCovariateBalanceViewer}
\alias{cohortMethodCovariateBalanceViewer}
\title{The module viewer for rendering the PLE covariate balance analysis}
\usage{
cohortMethodCovariateBalanceViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method covariate balance results
}
\description{
The module viewer for rendering the PLE covariate balance analysis
}
