% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-forestPlot.R
\name{cohortMethodForestPlotServer}
\alias{cohortMethodForestPlotServer}
\title{The module server for rendering the PLE multiple results forest plot}
\usage{
cohortMethodForestPlotServer(
  id,
  connectionHandler,
  selectedRow,
  inputParams,
  metaAnalysisDbIds = NULL,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{connection}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}

\item{metaAnalysisDbIds}{metaAnalysisDbIds}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
the PLE forest plot content server
}
\description{
The module server for rendering the PLE multiple results forest plot
}
