% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-kaplainMeier.R
\name{cohortMethodKaplanMeierViewer}
\alias{cohortMethodKaplanMeierViewer}
\title{The module viewer for rendering the PLE Kaplan Meier curve}
\usage{
cohortMethodKaplanMeierViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The module viewer for Kaplan Meier objects
}
\description{
The module viewer for rendering the PLE Kaplan Meier curve
}
