% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-kaplainMeier.R
\name{cohortMethodKaplanMeierViewer}
\alias{cohortMethodKaplanMeierViewer}
\title{The module viewer for rendering the PLE Kaplan Meier curve}
\usage{
cohortMethodKaplanMeierViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The module viewer for Kaplan Meier objects
}
\description{
The module viewer for rendering the PLE Kaplan Meier curve
}
\seealso{
Other {Estimation}: 
\code{\link{cohortMethodAttritionServer}()},
\code{\link{cohortMethodAttritionViewer}()},
\code{\link{cohortMethodCovariateBalanceServer}()},
\code{\link{cohortMethodCovariateBalanceViewer}()},
\code{\link{cohortMethodKaplanMeierServer}()},
\code{\link{cohortMethodPopulationCharacteristicsServer}()},
\code{\link{cohortMethodPopulationCharacteristicsViewer}()},
\code{\link{cohortMethodPowerServer}()},
\code{\link{cohortMethodPowerViewer}()},
\code{\link{cohortMethodPropensityModelServer}()},
\code{\link{cohortMethodPropensityModelViewer}()},
\code{\link{cohortMethodPropensityScoreDistServer}()},
\code{\link{cohortMethodPropensityScoreDistViewer}()},
\code{\link{cohortMethodSystematicErrorServer}()},
\code{\link{cohortMethodSystematicErrorViewer}()},
\code{\link{estimationHelperFile}()},
\code{\link{estimationServer}()},
\code{\link{estimationViewer}()}
}
\concept{{Estimation}}
