% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/estimation-cohort-method-populationCharacteristics.R
\name{cohortMethodPopulationCharacteristicsViewer}
\alias{cohortMethodPopulationCharacteristicsViewer}
\title{The module viewer for rendering the PLE population characteristics}
\usage{
cohortMethodPopulationCharacteristicsViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method population characteristics objects
}
\description{
The module viewer for rendering the PLE population characteristics
}
\seealso{
Other {Estimation}: 
\code{\link{cohortMethodAttritionServer}()},
\code{\link{cohortMethodAttritionViewer}()},
\code{\link{cohortMethodCovariateBalanceServer}()},
\code{\link{cohortMethodCovariateBalanceViewer}()},
\code{\link{cohortMethodKaplanMeierServer}()},
\code{\link{cohortMethodKaplanMeierViewer}()},
\code{\link{cohortMethodPopulationCharacteristicsServer}()},
\code{\link{cohortMethodPowerServer}()},
\code{\link{cohortMethodPowerViewer}()},
\code{\link{cohortMethodPropensityModelServer}()},
\code{\link{cohortMethodPropensityModelViewer}()},
\code{\link{cohortMethodPropensityScoreDistServer}()},
\code{\link{cohortMethodPropensityScoreDistViewer}()},
\code{\link{cohortMethodSystematicErrorServer}()},
\code{\link{cohortMethodSystematicErrorViewer}()},
\code{\link{estimationHelperFile}()},
\code{\link{estimationServer}()},
\code{\link{estimationViewer}()}
}
\concept{{Estimation}}
