% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/estimation-cohort-method-populationCharacteristics.R
\name{cohortMethodPopulationCharacteristicsViewer}
\alias{cohortMethodPopulationCharacteristicsViewer}
\title{The module viewer for rendering the PLE population characteristics}
\usage{
cohortMethodPopulationCharacteristicsViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method population characteristics objects
}
\description{
The module viewer for rendering the PLE population characteristics
}
