% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-power.R
\name{cohortMethodPowerServer}
\alias{cohortMethodPowerServer}
\title{The module server for rendering the PLE power analysis results}
\usage{
cohortMethodPowerServer(
  id,
  selectedRow,
  inputParams,
  connectionHandler,
  resultsSchema,
  tablePrefix,
  metaAnalysisDbIds = NULL
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultsSchema}{the schema with the PLE results}

\item{tablePrefix}{tablePrefix}

\item{metaAnalysisDbIds}{metaAnalysisDbIds}
}
\value{
the PLE systematic error power server
}
\description{
The module server for rendering the PLE power analysis results
}
