% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-power.R
\name{cohortMethodPowerServer}
\alias{cohortMethodPowerServer}
\title{The module server for rendering the PLE power analysis results}
\usage{
cohortMethodPowerServer(
  id,
  selectedRow,
  connectionHandler,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
the PLE systematic error power server
}
\description{
The module server for rendering the PLE power analysis results
}
