% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-power.R
\name{cohortMethodPowerViewer}
\alias{cohortMethodPowerViewer}
\title{The module viewer for rendering the PLE power analysis}
\usage{
cohortMethodPowerViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method power calculation results
}
\description{
The module viewer for rendering the PLE power analysis
}
