% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-propensityModel.R
\name{cohortMethodPropensityModelServer}
\alias{cohortMethodPropensityModelServer}
\title{The module server for rendering the propensity score model}
\usage{
cohortMethodPropensityModelServer(
  id,
  selectedRow,
  connectionHandler,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
the PLE propensity score model
}
\description{
The module server for rendering the propensity score model
}
\seealso{
Other Estimation: 
\code{\link{cohortMethodAttritionServer}()},
\code{\link{cohortMethodAttritionViewer}()},
\code{\link{cohortMethodCovariateBalanceServer}()},
\code{\link{cohortMethodCovariateBalanceViewer}()},
\code{\link{cohortMethodKaplanMeierServer}()},
\code{\link{cohortMethodKaplanMeierViewer}()},
\code{\link{cohortMethodPopulationCharacteristicsServer}()},
\code{\link{cohortMethodPopulationCharacteristicsViewer}()},
\code{\link{cohortMethodPowerServer}()},
\code{\link{cohortMethodPowerViewer}()},
\code{\link{cohortMethodPropensityModelViewer}()},
\code{\link{cohortMethodPropensityScoreDistServer}()},
\code{\link{cohortMethodPropensityScoreDistViewer}()},
\code{\link{cohortMethodSystematicErrorServer}()},
\code{\link{cohortMethodSystematicErrorViewer}()},
\code{\link{estimationHelperFile}()},
\code{\link{estimationServer}()},
\code{\link{estimationViewer}()}
}
\concept{Estimation}
