% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-propensityModel.R
\name{cohortMethodPropensityModelServer}
\alias{cohortMethodPropensityModelServer}
\title{The module server for rendering the propensity score model}
\usage{
cohortMethodPropensityModelServer(
  id,
  selectedRow,
  inputParams,
  connectionHandler,
  resultsSchema,
  tablePrefix
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultsSchema}{the schema with the PLE results}

\item{tablePrefix}{tablePrefix}
}
\value{
the PLE propensity score model
}
\description{
The module server for rendering the propensity score model
}
