% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-propensityModel.R
\name{cohortMethodPropensityModelViewer}
\alias{cohortMethodPropensityModelViewer}
\title{The module viewer for rendering the PLE propensity score model covariates/coefficients}
\usage{
cohortMethodPropensityModelViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method propensity score model covariates/coefficients
}
\description{
The module viewer for rendering the PLE propensity score model covariates/coefficients
}
