% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-propensityScoreDistribution.R
\name{cohortMethodPropensityScoreDistServer}
\alias{cohortMethodPropensityScoreDistServer}
\title{The module server for rendering a PLE propensity score distribution}
\usage{
cohortMethodPropensityScoreDistServer(
  id,
  selectedRow,
  inputParams,
  connectionHandler,
  resultsSchema,
  tablePrefix,
  cohortTablePrefix,
  metaAnalysisDbIds = F
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultsSchema}{the schema with the PLE results}

\item{tablePrefix}{tablePrefix}

\item{cohortTablePrefix}{cohortTablePrefix}

\item{metaAnalysisDbIds}{metaAnalysisDbIds}
}
\value{
the PLE propensity score distribution content server
}
\description{
The module server for rendering a PLE propensity score distribution
}
