% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/estimation-cohort-method-propensityScoreDistribution.R
\name{cohortMethodPropensityScoreDistServer}
\alias{cohortMethodPropensityScoreDistServer}
\title{The module server for rendering a PLE propensity score distribution}
\usage{
cohortMethodPropensityScoreDistServer(
  id,
  selectedRow,
  connectionHandler,
  resultDatabaseSettings,
  metaAnalysisDbIds = F
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{connectionHandler}{the connection to the PLE results database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}

\item{metaAnalysisDbIds}{metaAnalysisDbIds}
}
\value{
the PLE propensity score distribution content server
}
\description{
The module server for rendering a PLE propensity score distribution
}
