% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/estimation-cohort-method-propensityScoreDistribution.R
\name{cohortMethodPropensityScoreDistViewer}
\alias{cohortMethodPropensityScoreDistViewer}
\title{The module viewer for rendering the propensity score distribution}
\usage{
cohortMethodPropensityScoreDistViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method propensity score distribution
}
\description{
The module viewer for rendering the propensity score distribution
}
\seealso{
Other {Estimation}: 
\code{\link{cohortMethodAttritionServer}()},
\code{\link{cohortMethodAttritionViewer}()},
\code{\link{cohortMethodCovariateBalanceServer}()},
\code{\link{cohortMethodCovariateBalanceViewer}()},
\code{\link{cohortMethodKaplanMeierServer}()},
\code{\link{cohortMethodKaplanMeierViewer}()},
\code{\link{cohortMethodPopulationCharacteristicsServer}()},
\code{\link{cohortMethodPopulationCharacteristicsViewer}()},
\code{\link{cohortMethodPowerServer}()},
\code{\link{cohortMethodPowerViewer}()},
\code{\link{cohortMethodPropensityModelServer}()},
\code{\link{cohortMethodPropensityModelViewer}()},
\code{\link{cohortMethodPropensityScoreDistServer}()},
\code{\link{cohortMethodSystematicErrorServer}()},
\code{\link{cohortMethodSystematicErrorViewer}()},
\code{\link{estimationHelperFile}()},
\code{\link{estimationServer}()},
\code{\link{estimationViewer}()}
}
\concept{{Estimation}}
