% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-method-propensityScoreDistribution.R
\name{cohortMethodPropensityScoreDistViewer}
\alias{cohortMethodPropensityScoreDistViewer}
\title{The module viewer for rendering the propensity score distribution}
\usage{
cohortMethodPropensityScoreDistViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method propensity score distribution
}
\description{
The module viewer for rendering the propensity score distribution
}
