% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-systematicError.R
\name{cohortMethodSystematicErrorServer}
\alias{cohortMethodSystematicErrorServer}
\title{The module server for rendering the systematic error objects}
\usage{
cohortMethodSystematicErrorServer(
  id,
  selectedRow,
  connectionHandler,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{connectionHandler}{the connection handler to the result databases}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
the PLE systematic error content server
}
\description{
The module server for rendering the systematic error objects
}
\seealso{
Other {Estimation}: 
\code{\link{cohortMethodAttritionServer}()},
\code{\link{cohortMethodAttritionViewer}()},
\code{\link{cohortMethodCovariateBalanceServer}()},
\code{\link{cohortMethodCovariateBalanceViewer}()},
\code{\link{cohortMethodKaplanMeierServer}()},
\code{\link{cohortMethodKaplanMeierViewer}()},
\code{\link{cohortMethodPopulationCharacteristicsServer}()},
\code{\link{cohortMethodPopulationCharacteristicsViewer}()},
\code{\link{cohortMethodPowerServer}()},
\code{\link{cohortMethodPowerViewer}()},
\code{\link{cohortMethodPropensityModelServer}()},
\code{\link{cohortMethodPropensityModelViewer}()},
\code{\link{cohortMethodPropensityScoreDistServer}()},
\code{\link{cohortMethodPropensityScoreDistViewer}()},
\code{\link{cohortMethodSystematicErrorViewer}()},
\code{\link{estimationHelperFile}()},
\code{\link{estimationServer}()},
\code{\link{estimationViewer}()}
}
\concept{{Estimation}}
