% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-systematicError.R
\name{cohortMethodSystematicErrorServer}
\alias{cohortMethodSystematicErrorServer}
\title{The module server for rendering the systematic error objects}
\usage{
cohortMethodSystematicErrorServer(
  id,
  selectedRow,
  connectionHandler,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{connectionHandler}{the connection handler to the result databases}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
the PLE systematic error content server
}
\description{
The module server for rendering the systematic error objects
}
