% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-cohort-method-systematicError.R
\name{cohortMethodSystematicErrorViewer}
\alias{cohortMethodSystematicErrorViewer}
\title{The module viewer for rendering the PLE systematic error objects}
\usage{
cohortMethodSystematicErrorViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the cohort method systematic error module
}
\description{
The module viewer for rendering the PLE systematic error objects
}
