% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-cohort-overlap.R
\name{cohortOverlapView}
\alias{cohortOverlapView}
\title{Cohort Overlap View}
\usage{
cohortOverlapView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("cohortOverlap") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
