% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-compareCharacterization.R
\name{compareCohortCharacterizationView}
\alias{compareCohortCharacterizationView}
\title{compare characterization view}
\usage{
compareCohortCharacterizationView(
  id,
  title = "Compare cohort characterization"
)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("compareCohortCharacterization") inside diagnosticsExplorer module}

\item{title}{Optional string title field}
}
\description{
Use for customizing UI
}
\seealso{
Other CohortDiagnostics: 
\code{\link{cohortCountsModule}()},
\code{\link{cohortCountsView}()},
\code{\link{cohortDefinitionsModule}()},
\code{\link{cohortDefinitionsView}()},
\code{\link{cohortDiagCharacterizationView}()},
\code{\link{cohortDiagnosticsHelperFile}()},
\code{\link{cohortDiagnosticsServer}()},
\code{\link{cohortDiagnosticsView}()},
\code{\link{cohortOverlapView}()},
\code{\link{conceptsInDataSourceView}()},
\code{\link{createCdDatabaseDataSource}()},
\code{\link{databaseInformationView}()},
\code{\link{getCirceRenderedExpression}()},
\code{\link{getEnabledCdReports}()},
\code{\link{incidenceRatesView}()},
\code{\link{inclusionRulesView}()},
\code{\link{indexEventBreakdownView}()},
\code{\link{orpahanConceptsView}()},
\code{\link{timeDistributionsView}()},
\code{\link{visitContextView}()}
}
\concept{CohortDiagnostics}
