% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-compareCharacterization.R
\name{compareCohortCharacterizationView}
\alias{compareCohortCharacterizationView}
\title{compare characterization view}
\usage{
compareCohortCharacterizationView(
  id,
  title = "Compare cohort characterization",
  ...
)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("compareCohortCharacterization") inside diagnosticsExplorer module}

\item{title}{Optional string title field}
}
\description{
Use for customizing UI
}
