% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-conceptsInDataSource.R
\name{conceptsInDataSourceView}
\alias{conceptsInDataSourceView}
\title{concepts In DataSource View}
\usage{
conceptsInDataSourceView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("conceptsInDataSource") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
