% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covariateModule.R
\name{covariateServer}
\alias{covariateServer}
\title{The server for creating FeatureExtraction covariate settings}
\usage{
covariateServer(input, output, session, cohortReactive, validation = F)
}
\arguments{
\item{input}{Standard for shiny modules}

\item{output}{Standard for shiny modules}

\item{session}{Standard for shiny modules}

\item{cohortReactive}{The list of cohort in an ATLAS webApi}

\item{validation}{Whether the covariates are for development or validation}
}
\description{
The server for creating FeatureExtraction covariate settings
}
\examples{
\dontrun{
webApi <- callModule(webApiServer, 'webApiMain')
cohortReactive <- callModule(extractCohortsServer, 'extractCohort1',
                            webApi = webApi)
covList <- callModule(covariateServer, 'covariate1', cohortReactive)                            
}
}
