% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-actions.R
\name{createActionButton}
\alias{createActionButton}
\title{Create Action Button Definition}
\usage{
createActionButton(
  actionType,
  buttonIcon = "play",
  hoverText = NULL,
  buttonClass = "btn btn-default btn-xs",
  buttonStyle = NULL,
  buttonLabel = NULL
)
}
\arguments{
\item{actionType}{string, action identifier emitted by the button click event.}

\item{buttonIcon}{string, Font Awesome icon name (for shiny::icon). Use "" to hide the icon.}

\item{hoverText}{string, button tooltip shown on hover.}

\item{buttonClass}{string, CSS classes applied to the button element.}

\item{buttonStyle}{string, inline CSS style applied to the button element. Uses a compact pill style when NULL.}

\item{buttonLabel}{string, button text shown to the user.}
}
\value{
A list with action button settings.
}
\description{
Build an action button configuration for resultTableServer(addActions = ...).
}
\seealso{
Other Utils:
\code{\link[=actionButtonStyleInfo]{actionButtonStyleInfo()}},
\code{\link[=actionButtonStyleWarning]{actionButtonStyleWarning()}},
\code{\link[=datasourcesHelperFile]{datasourcesHelperFile()}},
\code{\link[=datasourcesServer]{datasourcesServer()}},
\code{\link[=datasourcesViewer]{datasourcesViewer()}},
\code{\link[=getLogoImage]{getLogoImage()}},
\code{\link[=resultTableViewer]{resultTableViewer()}}
}
\concept{Utils}
