% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main.R
\name{createCdDatabaseDataSource}
\alias{createCdDatabaseDataSource}
\title{Create a CD data source from a database}
\usage{
createCdDatabaseDataSource(
  connectionHandler,
  resultsDatabaseSchema,
  vocabularyDatabaseSchema = resultsDatabaseSchema,
  tablePrefix = "",
  cohortTableName = "cohort",
  databaseTableName = "database",
  dataModelSpecificationsPath = system.file("cohort-diagnostics-ref",
    "resultsDataModelSpecification.csv", package = utils::packageName())
)
}
\arguments{
\item{connectionHandler}{An instance of a ResultModelManager::connectionHander - manages a connection to a database.}

\item{resultsDatabaseSchema}{The schema containing the results tables in the database.}

\item{vocabularyDatabaseSchema}{The schema containing the vocabulary tables in the database. If not provided, defaults to `resultsDatabaseSchema`.}

\item{tablePrefix}{An optional prefix to add to the table names.}

\item{cohortTableName}{The name of the cohort table in the database.}

\item{databaseTableName}{The name of the database table in the database.}

\item{dataModelSpecificationsPath}{The path to a file containing specifications for the data model used by the database.}
}
\value{
An object of class `CdDataSource`.
}
\description{
use this to create an interface to cohort diagnostics results data
NOTE: I think this would make a good R6 class for other objects in this package so you could query them outside of
a shiny app. E.g. if you wanted to make a custom R markdown template
}
\examples{
# create a CD data source from a database
createCdDatabaseDataSource(connectionHandler, resultsDatabaseSchema)
}
