% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main.R
\name{createCdDatabaseDataSource}
\alias{createCdDatabaseDataSource}
\title{Create a CD data source from a database}
\usage{
createCdDatabaseDataSource(
  connectionHandler,
  resultDatabaseSettings,
  dataModelSpecificationsPath = system.file("cohort-diagnostics-ref",
    "resultsDataModelSpecification.csv", package = utils::packageName()),
  dataMigrationsRef = system.file("cohort-diagnostics-ref", "migrations.csv", package =
    utils::packageName()),
  displayProgress = FALSE
)
}
\arguments{
\item{connectionHandler}{An instance of a ResultModelManager::connectionHander - manages a connection to a database.}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}

\item{dataModelSpecificationsPath}{The path to a file containing specifications for the data model used by the database.}

\item{dataMigrationsRef}{The path to a file listing all migrations for the data model that should have been applied}

\item{displayProgress}{display a progress messaage (can only be used inside a shiny reactive context)}
}
\value{
An object of class `CdDataSource`.
}
\description{
use this to create an interface to cohort diagnostics results data
NOTE: I think this would make a good R6 class for other objects in this package so you could query them outside of
a shiny app. E.g. if you wanted to make a custom R markdown template
}
