% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-componentsCreateCustomColDefList.R
\name{createCustomColDefList}
\alias{createCustomColDefList}
\title{Creating a list of custom column definitions for use in reactables}
\usage{
createCustomColDefList(
  rawColNames,
  niceColNames = NULL,
  tooltipText = NULL,
  case = NULL,
  customColDefOptions = NULL
)
}
\arguments{
\item{rawColNames}{The raw column names taken directly from the source
data table that are to be overwritten in the reactable}

\item{niceColNames}{The formatted column names that will appear as-specified in 
the reactable}

\item{tooltipText}{The text to be displayed in a toolTip when hovering over the 
column in the reactable}

\item{case}{Optional argument to convert raw column names to snake or camel case. Defaults to NULL and preserves
whatever raw column names are passed in}

\item{customColDefOptions}{A list of lists, where the inner lists are any custom options from
reactable::colDef for each column}
}
\value{
A named list of reactable::colDef objects
}
\description{
Creating a list of custom column definitions for use in reactables
}
