% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-databaseInformation.R
\name{databaseInformationView}
\alias{databaseInformationView}
\title{database Information View}
\usage{
databaseInformationView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("databaseInformation") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
