% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-diagnostic-drill.R
\name{dataDiagnosticDrillServer}
\alias{dataDiagnosticDrillServer}
\title{The module server for exploring prediction summary results}
\usage{
dataDiagnosticDrillServer(id, connectionHandler, resultDatabaseSettings)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
The server to the summary module
}
\description{
The module server for exploring prediction summary results
}
\details{
The user specifies the id for the module
}
