% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-diagnostic-drill.R
\name{dataDiagnosticDrillServer}
\alias{dataDiagnosticDrillServer}
\title{The module server for exploring prediction summary results}
\usage{
dataDiagnosticDrillServer(id, con, mySchema, targetDialect, myTableAppend)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{con}{the connection to the prediction result database}

\item{mySchema}{the database schema for the model results}

\item{targetDialect}{the database management system for the model results}

\item{myTableAppend}{a string that appends the tables in the result schema}
}
\value{
The server to the summary module
}
\description{
The module server for exploring prediction summary results
}
\details{
The user specifies the id for the module
}
