% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-diagnostic-drill.R
\name{dataDiagnosticDrillViewer}
\alias{dataDiagnosticDrillViewer}
\title{The module viewer for exploring data-diagnostic results in more detail}
\usage{
dataDiagnosticDrillViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the summary module
}
\description{
The module viewer for exploring data-diagnostic results in more detail
}
\details{
The user specifies the id for the module
}
\seealso{
Other DataDiagnostics: 
\code{\link{dataDiagnosticDrillServer}()},
\code{\link{dataDiagnosticHelperFile}()},
\code{\link{dataDiagnosticServer}()},
\code{\link{dataDiagnosticSummaryServer}()},
\code{\link{dataDiagnosticSummaryViewer}()},
\code{\link{dataDiagnosticViewer}()}
}
\concept{DataDiagnostics}
