% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-diagnostic-main.R
\name{dataDiagnosticHelperFile}
\alias{dataDiagnosticHelperFile}
\title{The location of the data-diagnostic module helper file}
\usage{
dataDiagnosticHelperFile()
}
\value{
string location of the data-diagnostic helper file
}
\description{
The location of the data-diagnostic module helper file
}
\details{
Returns the location of the data-diagnostic helper file
}
\seealso{
Other {DataDiagnostics}: 
\code{\link{dataDiagnosticDrillServer}()},
\code{\link{dataDiagnosticDrillViewer}()},
\code{\link{dataDiagnosticServer}()},
\code{\link{dataDiagnosticSummaryServer}()},
\code{\link{dataDiagnosticSummaryViewer}()},
\code{\link{dataDiagnosticViewer}()}
}
\concept{{DataDiagnostics}}
