% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-diagnostic-main.R
\name{dataDiagnosticServer}
\alias{dataDiagnosticServer}
\title{The module server for exploring data-diagnostic}
\usage{
dataDiagnosticServer(id = "dataDiag", resultDatabaseSettings = list(port = 1))
}
\arguments{
\item{id}{the unique reference id for the module}

\item{resultDatabaseSettings}{a list containing the data-diagnostic result schema and connection details}
}
\value{
The server for the data-diagnostic module
}
\description{
The module server for exploring data-diagnostic
}
\details{
The user specifies the id for the module
}
