% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-diagnostic-summary.R
\name{dataDiagnosticSummaryViewer}
\alias{dataDiagnosticSummaryViewer}
\title{The module viewer for exploring data-diagnostic summary results}
\usage{
dataDiagnosticSummaryViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the summary module
}
\description{
The module viewer for exploring data-diagnostic summary results
}
\details{
The user specifies the id for the module
}
\seealso{
Other {DataDiagnostics}: 
\code{\link{dataDiagnosticDrillServer}()},
\code{\link{dataDiagnosticDrillViewer}()},
\code{\link{dataDiagnosticHelperFile}()},
\code{\link{dataDiagnosticServer}()},
\code{\link{dataDiagnosticSummaryServer}()},
\code{\link{dataDiagnosticViewer}()}
}
\concept{{DataDiagnostics}}
