% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-diagnostic-main.R
\name{dataDiagnosticViewer}
\alias{dataDiagnosticViewer}
\title{The module viewer for exploring data-diagnostic}
\usage{
dataDiagnosticViewer(id = "dataDiag")
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the data-diagnostic viewer module
}
\description{
The module viewer for exploring data-diagnostic
}
\details{
The user specifies the id for the module
}
\seealso{
Other {DataDiagnostics}: 
\code{\link{dataDiagnosticDrillServer}()},
\code{\link{dataDiagnosticDrillViewer}()},
\code{\link{dataDiagnosticHelperFile}()},
\code{\link{dataDiagnosticServer}()},
\code{\link{dataDiagnosticSummaryServer}()},
\code{\link{dataDiagnosticSummaryViewer}()}
}
\concept{{DataDiagnostics}}
