% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasources-main.R
\name{datasourcesHelperFile}
\alias{datasourcesHelperFile}
\title{Define the helper file for the module}
\usage{
datasourcesHelperFile()
}
\value{
The helper html file for the datasources module
}
\description{
Define the helper file for the module
}
\seealso{
Other {Utils}: 
\code{\link{createCustomColDefList}()},
\code{\link{datasourcesServer}()},
\code{\link{datasourcesViewer}()},
\code{\link{makeButtonLabel}()},
\code{\link{resultTableServer}()},
\code{\link{resultTableViewer}()}

Other {Utils}: 
\code{\link{createCustomColDefList}()},
\code{\link{datasourcesServer}()},
\code{\link{datasourcesViewer}()},
\code{\link{makeButtonLabel}()},
\code{\link{resultTableServer}()},
\code{\link{resultTableViewer}()}
}
\concept{{Utils}}
