% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasources-main.R
\name{datasourcesHelperFile}
\alias{datasourcesHelperFile}
\title{Define the helper file for the module}
\usage{
datasourcesHelperFile()
}
\value{
The helper html file for the datasources module
}
\description{
Define the helper file for the module
}
