% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasources-main.R
\name{datasourcesServer}
\alias{datasourcesServer}
\title{The server function for the datasources module}
\usage{
datasourcesServer(id, connectionHandler, resultDatabaseSettings)
}
\arguments{
\item{id}{The unique id for the datasources server namespace}

\item{connectionHandler}{A connection to the database with the results}

\item{resultDatabaseSettings}{A named list containing the cohort generator results database details (schema, table prefix)}
}
\value{
The server for the datasources module
}
\description{
The server function for the datasources module
}
