% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasources-main.R
\name{datasourcesViewer}
\alias{datasourcesViewer}
\title{The viewer function for hte datasources module}
\usage{
datasourcesViewer(id)
}
\arguments{
\item{id}{The unique id for the datasources viewer namespace}
}
\value{
The UI for the datasources module
}
\description{
The viewer function for hte datasources module
}
\seealso{
Other Utils:
\code{\link[=datasourcesHelperFile]{datasourcesHelperFile()}},
\code{\link[=datasourcesServer]{datasourcesServer()}},
\code{\link[=getLogoImage]{getLogoImage()}},
\code{\link[=resultTableViewer]{resultTableViewer()}}
}
\concept{Utils}
