% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-aggregateFeatures.R
\name{descriptionAggregateFeaturesServer}
\alias{descriptionAggregateFeaturesServer}
\title{The module server for exploring aggregate features results}
\usage{
descriptionAggregateFeaturesServer(
  id,
  con,
  mainPanelTab,
  schema,
  dbms,
  tablePrefix,
  tempEmulationSchema = NULL,
  cohortTablePrefix = "cg_",
  databaseTable = "DATABASE_META_DATA"
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{con}{the connection to the prediction result database}

\item{mainPanelTab}{the current tab}

\item{schema}{the database schema for the model results}

\item{dbms}{the database management system for the model results}

\item{tablePrefix}{a string that appends the tables in the result schema}

\item{tempEmulationSchema}{The temp schema (optional)}

\item{cohortTablePrefix}{a string that appends the COHORT_DEFINITION table in the result schema}

\item{databaseTable}{The database table name}
}
\value{
The server to the description aggregate features module
}
\description{
The module server for exploring aggregate features results
}
\details{
The user specifies the id for the module
}
