% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/description-main.R
\name{descriptionServer}
\alias{descriptionServer}
\title{The module server for exploring description studies}
\usage{
descriptionServer(
  id,
  connectionHandler,
  resultDatabaseSettings = list(port = 1)
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list containing the description result schema, dbms, tablePrefix, databaseTable and cohortTablePrefix}
}
\value{
The server for the description module
}
\description{
The module server for exploring description studies
}
\details{
The user specifies the id for the module
}
