% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-main.R
\name{estimationHelperFile}
\alias{estimationHelperFile}
\title{The location of the estimation module helper file}
\usage{
estimationHelperFile()
}
\value{
string location of the characterization helper file
}
\description{
The location of the estimation module helper file
}
\details{
Returns the location of the characterization helper file
}
