% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-kaplainMeier.R
\name{estimationKaplanMeierServer}
\alias{estimationKaplanMeierServer}
\title{The module server for rendering the Kaplan Meier curve}
\usage{
estimationKaplanMeierServer(
  id,
  selectedRow,
  inputParams,
  connection,
  resultsSchema,
  tablePrefix,
  cohortTablePrefix,
  databaseTable
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}

\item{connection}{the connection to the PLE results database}

\item{resultsSchema}{the schema with the PLE results}
}
\value{
the PLE Kaplain Meier content server
}
\description{
The module server for rendering the Kaplan Meier curve
}
