% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-main.R
\name{estimationServer}
\alias{estimationServer}
\title{The module server for the main estimation module}
\usage{
estimationServer(id, connectionHandler, resultDatabaseSettings)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a named list containing the PLE results database connection details}
}
\value{
the PLE results viewer main module server
}
\description{
The module server for the main estimation module
}
