% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-main.R
\name{estimationServer}
\alias{estimationServer}
\title{The module server for exploring estimation studies}
\usage{
estimationServer(
  id,
  connectionHandler,
  resultDatabaseSettings = list(port = 1)
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list containing the characterization result schema, dbms, tablePrefix, databaseTable and cgTablePrefix}
}
\value{
The server for the estimation module
}
\description{
The module server for exploring estimation studies
}
\details{
The user specifies the id for the module
}
