% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-subgroups.R
\name{estimationSubgroupsServer}
\alias{estimationSubgroupsServer}
\title{The module server for rendering the subgroup results}
\usage{
estimationSubgroupsServer(
  id,
  selectedRow,
  inputParams,
  exposureOfInterest,
  outcomeOfInterest,
  connectionHandler
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{selectedRow}{the selected row from the main results table}

\item{inputParams}{the selected study parameters of interest}

\item{exposureOfInterest}{exposureOfInterest}

\item{outcomeOfInterest}{outcomeOfInterest}

\item{connectionHandler}{connection}
}
\value{
the PLE subgroup results server
}
\description{
The module server for rendering the subgroup results
}
