% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-main.R
\name{estimationViewer}
\alias{estimationViewer}
\title{The module viewer for exploring characterization studies}
\usage{
estimationViewer(id = 1)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the characterization viewer module
}
\description{
The module viewer for exploring characterization studies
}
\details{
The user specifies the id for the module
}
