% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimation-main.R
\name{estimationViewer}
\alias{estimationViewer}
\title{The viewer of the main estimation module}
\usage{
estimationViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the estimation results viewer
}
\description{
The viewer of the main estimation module
}
