% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-definition.R
\name{getCirceRenderedExpression}
\alias{getCirceRenderedExpression}
\title{Returns list with circe generated documentation}
\usage{
getCirceRenderedExpression(
  cohortDefinition,
  cohortName = "Cohort Definition",
  includeConceptSets = FALSE
)
}
\arguments{
\item{cohortDefinition}{An R object (list) with a list representation of the cohort definition expression,
that may be converted to a cohort expression JSON using
jsonlite::toJSON(x = cohortDefinition, digits = 23, pretty = TRUE)}

\item{cohortName}{Name for the cohort definition}

\item{includeConceptSets}{Do you want to inclued concept set in the documentation}
}
\value{
list object
}
\description{
Returns list with circe generated documentation
}
\seealso{
Other CohortDiagnostics: 
\code{\link{cohortCountsModule}()},
\code{\link{cohortCountsView}()},
\code{\link{cohortDefinitionsModule}()},
\code{\link{cohortDefinitionsView}()},
\code{\link{cohortDiagCharacterizationView}()},
\code{\link{cohortDiagnosticsHelperFile}()},
\code{\link{cohortDiagnosticsServer}()},
\code{\link{cohortDiagnosticsView}()},
\code{\link{cohortOverlapView}()},
\code{\link{compareCohortCharacterizationView}()},
\code{\link{conceptsInDataSourceView}()},
\code{\link{createCdDatabaseDataSource}()},
\code{\link{databaseInformationView}()},
\code{\link{getEnabledCdReports}()},
\code{\link{incidenceRatesView}()},
\code{\link{inclusionRulesView}()},
\code{\link{indexEventBreakdownView}()},
\code{\link{orpahanConceptsView}()},
\code{\link{timeDistributionsView}()},
\code{\link{visitContextView}()}
}
\concept{CohortDiagnostics}
