% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main.R
\name{getEnabledCdReports}
\alias{getEnabledCdReports}
\title{Get enable cd reports from available data}
\usage{
getEnabledCdReports(dataSource)
}
\arguments{
\item{dataSource}{Cohort diagnostics data source}
}
\description{
Get enable cd reports from available data
}
