% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-main.R
\name{getEnabledCdReports}
\alias{getEnabledCdReports}
\title{Get enable cd reports from available data}
\usage{
getEnabledCdReports(dataSource)
}
\arguments{
\item{dataSource}{Cohort diagnostics data source}
}
\description{
Get enable cd reports from available data
}
\seealso{
Other {CohortDiagnostics}: 
\code{\link{cohortCountsModule}()},
\code{\link{cohortCountsView}()},
\code{\link{cohortDefinitionsModule}()},
\code{\link{cohortDefinitionsView}()},
\code{\link{cohortDiagCharacterizationView}()},
\code{\link{cohortDiagnosticsHelperFile}()},
\code{\link{cohortDiagnosticsServer}()},
\code{\link{cohortDiagnosticsView}()},
\code{\link{cohortOverlapView}()},
\code{\link{compareCohortCharacterizationView}()},
\code{\link{conceptsInDataSourceView}()},
\code{\link{createCdDatabaseDataSource}()},
\code{\link{databaseInformationView}()},
\code{\link{getCirceRenderedExpression}()},
\code{\link{incidenceRatesView}()},
\code{\link{inclusionRulesView}()},
\code{\link{indexEventBreakdownView}()},
\code{\link{orpahanConceptsView}()},
\code{\link{timeDistributionsView}()},
\code{\link{visitContextView}()}
}
\concept{{CohortDiagnostics}}
