% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-logo.R
\name{getLogoImage}
\alias{getLogoImage}
\title{The location of the OHDSI logo}
\usage{
getLogoImage()
}
\value{
string location of the OHDSI logo
}
\description{
The location of the OHDSI logo
}
\details{
Returns the location of the OHDSI logo
}
\seealso{
Other Utils:
\code{\link[=actionButtonStyleInfo]{actionButtonStyleInfo()}},
\code{\link[=actionButtonStyleWarning]{actionButtonStyleWarning()}},
\code{\link[=createActionButton]{createActionButton()}},
\code{\link[=datasourcesHelperFile]{datasourcesHelperFile()}},
\code{\link[=datasourcesServer]{datasourcesServer()}},
\code{\link[=datasourcesViewer]{datasourcesViewer()}},
\code{\link[=resultTableViewer]{resultTableViewer()}}
}
\concept{Utils}
