% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers-logo.R
\name{getLogoImage}
\alias{getLogoImage}
\title{The location of the OHDSI logo}
\usage{
getLogoImage()
}
\value{
string location of the OHDSI logo
}
\description{
The location of the OHDSI logo
}
\details{
Returns the location of the OHDSI logo
}
