% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home-main.R
\name{homeHelperFile}
\alias{homeHelperFile}
\title{The location of the home module helper file}
\usage{
homeHelperFile()
}
\value{
string location of the home helper file
}
\description{
The location of the home module helper file
}
\details{
Returns the location of the home helper file
}
