% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/home-main.R
\name{homeViewer}
\alias{homeViewer}
\title{The module viewer for exploring home}
\usage{
homeViewer(id = 1)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the home viewer module
}
\description{
The module viewer for exploring home
}
\details{
The user specifies the id for the module
}
