% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-incidenceRates.R
\name{incidenceRatesView}
\alias{incidenceRatesView}
\title{incidence Rates View}
\usage{
incidenceRatesView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("incidenceRates") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
\seealso{
Other {CohortDiagnostics}: 
\code{\link{cohortCountsModule}()},
\code{\link{cohortCountsView}()},
\code{\link{cohortDefinitionsModule}()},
\code{\link{cohortDefinitionsView}()},
\code{\link{cohortDiagCharacterizationView}()},
\code{\link{cohortDiagnosticsHelperFile}()},
\code{\link{cohortDiagnosticsServer}()},
\code{\link{cohortDiagnosticsView}()},
\code{\link{cohortOverlapView}()},
\code{\link{compareCohortCharacterizationView}()},
\code{\link{conceptsInDataSourceView}()},
\code{\link{createCdDatabaseDataSource}()},
\code{\link{databaseInformationView}()},
\code{\link{getCirceRenderedExpression}()},
\code{\link{getEnabledCdReports}()},
\code{\link{inclusionRulesView}()},
\code{\link{indexEventBreakdownView}()},
\code{\link{orpahanConceptsView}()},
\code{\link{timeDistributionsView}()},
\code{\link{visitContextView}()}
}
\concept{{CohortDiagnostics}}
