% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-incidenceRates.R
\name{incidenceRatesView}
\alias{incidenceRatesView}
\title{incidence Rates View}
\usage{
incidenceRatesView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("incidenceRates") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
