% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-inclusionRules.R
\name{inclusionRulesView}
\alias{inclusionRulesView}
\title{inclusion Rules View}
\usage{
inclusionRulesView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("inclusionRules") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
