% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-indexEventBreakdown.R
\name{indexEventBreakdownView}
\alias{indexEventBreakdownView}
\title{Index event breakdown view}
\usage{
indexEventBreakdownView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("indexEvents") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
