% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-largeTableViewer.R
\name{largeTableView}
\alias{largeTableView}
\title{Large Table Component Viewer}
\usage{
largeTableView(id, pageSizeChoices = c(10, 20, 50, 100))
}
\arguments{
\item{id}{Shiny module id. Must match largeTableServer}

\item{pageSizeChoices}{numeric selection options for pages}
}
\description{
Componenet for results sets with many thousands of rows
More limited than other table components in terms of automatic handling of search and
filtering but will allow responsive apps
}
