% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OhdsiShinyModules.R
\docType{package}
\name{OhdsiShinyModules}
\alias{OhdsiShinyModules-package}
\alias{OhdsiShinyModules}
\title{OhdsiShinyModules}
\description{
A selection of shiny modules for exploring standardized OHDSI results
}
