% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-orphanConcepts.R
\name{orpahanConceptsView}
\alias{orpahanConceptsView}
\title{Orphan Concepts View}
\usage{
orpahanConceptsView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("orphanConcepts") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
