% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-diagnostics.R
\name{patientLevelPredictionDiagnosticsServer}
\alias{patientLevelPredictionDiagnosticsServer}
\title{The module server for exploring prediction diagnostic results}
\usage{
patientLevelPredictionDiagnosticsServer(
  id,
  modelDesignId,
  connectionHandler,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{modelDesignId}{the unique id for the model design}

\item{connectionHandler}{the connection to the prediction result database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
The server to the prediction diagnostic module
}
\description{
The module server for exploring prediction diagnostic results
}
\details{
The user specifies the id for the module
}
