% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-discrimination.R
\name{patientLevelPredictionDiscriminationViewer}
\alias{patientLevelPredictionDiscriminationViewer}
\title{The module viewer for exploring prediction model discrimination results}
\usage{
patientLevelPredictionDiscriminationViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the model discrimination results module
}
\description{
The module viewer for exploring prediction model discrimination results
}
\details{
The user specifies the id for the module
}
\seealso{
Other {PatientLevelPrediction}: 
\code{\link{patientLevelPredictionCalibrationServer}()},
\code{\link{patientLevelPredictionCalibrationViewer}()},
\code{\link{patientLevelPredictionCovariateSummaryServer}()},
\code{\link{patientLevelPredictionCovariateSummaryViewer}()},
\code{\link{patientLevelPredictionCutoffServer}()},
\code{\link{patientLevelPredictionCutoffViewer}()},
\code{\link{patientLevelPredictionDesignSummaryServer}()},
\code{\link{patientLevelPredictionDesignSummaryViewer}()},
\code{\link{patientLevelPredictionDiagnosticsServer}()},
\code{\link{patientLevelPredictionDiagnosticsViewer}()},
\code{\link{patientLevelPredictionDiscriminationServer}()},
\code{\link{patientLevelPredictionHelperFile}()},
\code{\link{patientLevelPredictionModelSummaryServer}()},
\code{\link{patientLevelPredictionModelSummaryViewer}()},
\code{\link{patientLevelPredictionNbServer}()},
\code{\link{patientLevelPredictionNbViewer}()},
\code{\link{patientLevelPredictionServer}()},
\code{\link{patientLevelPredictionSettingsServer}()},
\code{\link{patientLevelPredictionSettingsViewer}()},
\code{\link{patientLevelPredictionValidationServer}()},
\code{\link{patientLevelPredictionValidationViewer}()},
\code{\link{patientLevelPredictionViewer}()}
}
\concept{{PatientLevelPrediction}}
