% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-heatmap.R
\name{patientLevelPredictionHeatmapViewer}
\alias{patientLevelPredictionHeatmapViewer}
\title{The module viewer for exploring prediction model discrimination results}
\usage{
patientLevelPredictionHeatmapViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the model discrimination results module
}
\description{
The module viewer for exploring prediction model discrimination results
}
\details{
The user specifies the id for the module
}
\seealso{
Other PatientLevelPrediction: 
\code{\link{patientLevelPredictionHeatmapServer}()},
\code{\link{patientLevelPredictionHelperFile}()},
\code{\link{patientLevelPredictionModelServer}()},
\code{\link{patientLevelPredictionModelViewer}()},
\code{\link{patientLevelPredictionPlotServer}()},
\code{\link{patientLevelPredictionPlotViewer}()},
\code{\link{patientLevelPredictionServer}()},
\code{\link{patientLevelPredictionViewer}()}
}
\concept{PatientLevelPrediction}
