% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-main.R
\name{patientLevelPredictionHelperFile}
\alias{patientLevelPredictionHelperFile}
\title{The location of the prediction module helper file}
\usage{
patientLevelPredictionHelperFile()
}
\value{
string location of the prediction helper file
}
\description{
The location of the prediction module helper file
}
\details{
Returns the location of the prediction helper file
}
