% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-model.R
\name{patientLevelPredictionModelServer}
\alias{patientLevelPredictionModelServer}
\title{The module server for exploring prediction models}
\usage{
patientLevelPredictionModelServer(
  id,
  performances,
  performanceRowIds,
  connectionHandler,
  resultDatabaseSettings
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{performances}{a reactive val with the performance data.frame}

\item{performanceRowIds}{a vector of selected performance rows}

\item{connectionHandler}{the connection to the prediction result database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}
}
\value{
The server to the model module
}
\description{
The module server for exploring prediction models
}
\details{
The user specifies the id for the module
}
\seealso{
Other PatientLevelPrediction: 
\code{\link{patientLevelPredictionHelperFile}()},
\code{\link{patientLevelPredictionModelViewer}()},
\code{\link{patientLevelPredictionPlotServer}()},
\code{\link{patientLevelPredictionPlotViewer}()},
\code{\link{patientLevelPredictionServer}()},
\code{\link{patientLevelPredictionViewer}()}
}
\concept{PatientLevelPrediction}
