% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-modelSummary.R
\name{patientLevelPredictionModelSummaryServer}
\alias{patientLevelPredictionModelSummaryServer}
\title{The module server for exploring prediction summary results}
\usage{
patientLevelPredictionModelSummaryServer(
  id,
  connectionHandler,
  resultDatabaseSettings,
  modelDesignId
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{resultDatabaseSettings}{a list containing the result schema and prefixes}

\item{modelDesignId}{a reactable id specifying the prediction model design identifier}
}
\value{
The server to the summary module
}
\description{
The module server for exploring prediction summary results
}
\details{
The user specifies the id for the module
}
\seealso{
Other {PatientLevelPrediction}: 
\code{\link{patientLevelPredictionCalibrationServer}()},
\code{\link{patientLevelPredictionCalibrationViewer}()},
\code{\link{patientLevelPredictionCovariateSummaryServer}()},
\code{\link{patientLevelPredictionCovariateSummaryViewer}()},
\code{\link{patientLevelPredictionCutoffServer}()},
\code{\link{patientLevelPredictionCutoffViewer}()},
\code{\link{patientLevelPredictionDesignSummaryServer}()},
\code{\link{patientLevelPredictionDesignSummaryViewer}()},
\code{\link{patientLevelPredictionDiagnosticsServer}()},
\code{\link{patientLevelPredictionDiagnosticsViewer}()},
\code{\link{patientLevelPredictionDiscriminationServer}()},
\code{\link{patientLevelPredictionDiscriminationViewer}()},
\code{\link{patientLevelPredictionHelperFile}()},
\code{\link{patientLevelPredictionModelSummaryViewer}()},
\code{\link{patientLevelPredictionNbServer}()},
\code{\link{patientLevelPredictionNbViewer}()},
\code{\link{patientLevelPredictionServer}()},
\code{\link{patientLevelPredictionSettingsServer}()},
\code{\link{patientLevelPredictionSettingsViewer}()},
\code{\link{patientLevelPredictionValidationServer}()},
\code{\link{patientLevelPredictionValidationViewer}()},
\code{\link{patientLevelPredictionViewer}()}
}
\concept{{PatientLevelPrediction}}
