% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-model.R
\name{patientLevelPredictionModelViewer}
\alias{patientLevelPredictionModelViewer}
\title{The module viewer for exploring prediction models}
\usage{
patientLevelPredictionModelViewer(id)
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the model module
}
\description{
The module viewer for exploring prediction models
}
\details{
The user specifies the id for the module
}
\seealso{
Other PatientLevelPrediction: 
\code{\link{patientLevelPredictionHelperFile}()},
\code{\link{patientLevelPredictionModelServer}()},
\code{\link{patientLevelPredictionPlotServer}()},
\code{\link{patientLevelPredictionPlotViewer}()},
\code{\link{patientLevelPredictionServer}()},
\code{\link{patientLevelPredictionViewer}()}
}
\concept{PatientLevelPrediction}
