% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patient-level-prediction-main.R
\name{patientLevelPredictionServer}
\alias{patientLevelPredictionServer}
\title{The module server for exploring PatientLevelPrediction}
\usage{
patientLevelPredictionServer(
  id,
  connectionHandler,
  resultDatabaseSettings = list(port = 1)
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list containing the prediction result schema and connection details}
}
\value{
The server for the PatientLevelPrediction module
}
\description{
The module server for exploring PatientLevelPrediction
}
\details{
The user specifies the id for the module
}
\seealso{
Other PatientLevelPrediction: 
\code{\link{patientLevelPredictionHeatmapServer}()},
\code{\link{patientLevelPredictionHeatmapViewer}()},
\code{\link{patientLevelPredictionHelperFile}()},
\code{\link{patientLevelPredictionModelServer}()},
\code{\link{patientLevelPredictionModelViewer}()},
\code{\link{patientLevelPredictionPlotServer}()},
\code{\link{patientLevelPredictionPlotViewer}()},
\code{\link{patientLevelPredictionViewer}()}
}
\concept{PatientLevelPrediction}
