% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phevaluator-main.R
\name{phevaluatorServer}
\alias{phevaluatorServer}
\title{The module server for the main phevaluator module}
\usage{
phevaluatorServer(id, connectionHandler, resultDatabaseSettings)
}
\arguments{
\item{id}{The unique reference id for the module}

\item{connectionHandler}{A connection to the database with the results}

\item{resultDatabaseSettings}{A named list containing the cohort generator results database details (schema, table prefix)}
}
\value{
The phevaluator main module server
}
\description{
The module server for the main phevaluator module
}
\seealso{
Other PheValuator: 
\code{\link{phevaluatorHelperFile}()},
\code{\link{phevaluatorViewer}()}
}
\concept{PheValuator}
