% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phevaluator-main.R
\name{phevaluatorViewer}
\alias{phevaluatorViewer}
\title{The viewer of the phevaluator module}
\usage{
phevaluatorViewer(id)
}
\arguments{
\item{id}{The unique reference id for the module}
}
\value{
The user interface to the phevaluator results viewer
}
\description{
The viewer of the phevaluator module
}
