% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-designSummary.R
\name{predictionDesignSummaryServer}
\alias{predictionDesignSummaryServer}
\title{The module server for exploring prediction designs in the results}
\usage{
predictionDesignSummaryServer(
  id,
  connectionHandler,
  schema,
  plpTablePrefix,
  cohortTablePrefix
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{the connection to the prediction result database}

\item{schema}{the database schema for the model results}

\item{plpTablePrefix}{a string that appends the plp tables in the result schema}

\item{cohortTablePrefix}{a string that appends the cohort tables in the result schema}
}
\value{
The server to the prediction design module
}
\description{
The module server for exploring prediction designs in the results
}
\details{
The user specifies the id for the module
}
