% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-diagnostics.R
\name{predictionDiagnosticsServer}
\alias{predictionDiagnosticsServer}
\title{The module server for exploring prediction diagnostic results}
\usage{
predictionDiagnosticsServer(
  id,
  modelDesignId,
  mySchema,
  con,
  myTableAppend,
  targetDialect
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{modelDesignId}{the unique id for the model design}

\item{mySchema}{the database schema for the model results}

\item{con}{the connection to the prediction result database}

\item{myTableAppend}{a string that appends the tables in the result schema}

\item{targetDialect}{the database management system for the model results}
}
\value{
The server to the predcition diagnostic module
}
\description{
The module server for exploring prediction diagnostic results
}
\details{
The user specifies the id for the module
}
