% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-modelSummary.R
\name{predictionModelSummaryServer}
\alias{predictionModelSummaryServer}
\title{The module server for exploring prediction summary results}
\usage{
predictionModelSummaryServer(
  id,
  con,
  mySchema,
  targetDialect,
  myTableAppend,
  modelDesignId,
  databaseTableAppend
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{con}{the connection to the prediction result database}

\item{mySchema}{the database schema for the model results}

\item{targetDialect}{the database management system for the model results}

\item{myTableAppend}{a string that appends the tables in the result schema}

\item{modelDesignId}{a reactable id specifying the prediction model design identifier}

\item{databaseTableAppend}{a string that appends the database_meta_data table}
}
\value{
The server to the summary module
}
\description{
The module server for exploring prediction summary results
}
\details{
The user specifies the id for the module
}
