% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-main.R
\name{predictionServer}
\alias{predictionServer}
\title{The module server for exploring PatientLevelPrediction}
\usage{
predictionServer(
  id,
  connectionHandler,
  resultDatabaseSettings = list(port = 1)
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list containing the prediction result schema and connection details}
}
\value{
The server for the PatientLevelPrediction module
}
\description{
The module server for exploring PatientLevelPrediction
}
\details{
The user specifies the id for the module
}
