% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction-settings.R
\name{predictionSettingsServer}
\alias{predictionSettingsServer}
\title{The module server for exploring prediction settings}
\usage{
predictionSettingsServer(
  id,
  modelDesignId,
  developmentDatabaseId,
  performanceId,
  connectionHandler,
  inputSingleView,
  mySchema,
  myTableAppend,
  cohortTableAppend = myTableAppend
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{modelDesignId}{unique id for the model design}

\item{developmentDatabaseId}{unique id for the development database}

\item{performanceId}{unique id for the performance results}

\item{connectionHandler}{the connection to the prediction result database}

\item{inputSingleView}{the current tab}

\item{mySchema}{the database schema for the model results}

\item{myTableAppend}{a string that appends the tables in the result schema}

\item{cohortTableAppend}{a string that appends the cohort_definition table}
}
\value{
The server to the settings module
}
\description{
The module server for exploring prediction settings
}
\details{
The user specifies the id for the module
}
