% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heplers-componentsCreateCustomColDefList.R
\name{readColDefsFromJSON}
\alias{readColDefsFromJSON}
\title{Read reactable::colDefs froma  JSON file}
\usage{
readColDefsFromJSON(filename, path = "")
}
\arguments{
\item{filename}{The dname of the input JSON file}

\item{path}{The path to where the JSON is located}
}
\description{
Read reactable::colDefs froma  JSON file
}
