% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-main.R
\name{reportServer}
\alias{reportServer}
\title{The module server for the shiny app report module}
\usage{
reportServer(
  id = "reportModule",
  connectionHandler = NULL,
  resultDatabaseSettings = NULL,
  server = Sys.getenv("RESULTS_SERVER"),
  username = Sys.getenv("RESULTS_USER"),
  password = Sys.getenv("RESULTS_PASSWORD"),
  dbms = "postgresql"
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list containing the characterization result schema, dbms, tablePrefix, databaseTable and cgTablePrefix}

\item{server}{server for the connection to the results for quarto}

\item{username}{username for the connection to the results for quarto}

\item{password}{password for the connection to the results for quarto}

\item{dbms}{dbms for the connection to the results for quarto}
}
\value{
The server for the shiny app home
}
\description{
The module server for the shiny app report module
}
\details{
The user specifies the id for the module
}
