% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-main.R
\name{reportViewer}
\alias{reportViewer}
\title{The module viewer for the shiny app report module}
\usage{
reportViewer(id = "reportModule")
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the home page module
}
\description{
The module viewer for the shiny app report module
}
\details{
The user specifies the id for the module
}
\seealso{
Other {Report}: 
\code{\link{reportHelperFile}()},
\code{\link{reportServer}()}
}
\concept{{Report}}
