% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-data-viewer.R
\name{resultTableServer}
\alias{resultTableServer}
\title{Result Table Server}
\usage{
resultTableServer(
  id,
  df,
  colDefsInput,
  selectedCols = NULL,
  sortedCols = NULL,
  addActions = NULL,
  downloadedFileName = NULL
)
}
\arguments{
\item{id}{string, table id must match resultsTableViewer function}

\item{df}{reactive that returns a data frame}

\item{colDefsInput}{named list of reactable::colDefs}

\item{selectedCols}{string vector of columns the reactable should display to start by default. Defaults to ALL if not specified.}

\item{sortedCols}{string vector of columns the reactable should sort by by default. Defaults to no sort if not specified.}

\item{addActions}{add a button row selector column to the table to a column called 'actions'.
actions must be a column in df}

\item{downloadedFileName}{string, desired name of downloaded data file. can use the name from the module that is being used}

\item{elementId}{optional string vector of element Id name for custom dropdown filtering if present in the customColDef list. Defaults to NULL.}
}
\value{
shiny module server
}
\description{
Result Table Server
}
