% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-data-viewer.R
\name{resultTableServer}
\alias{resultTableServer}
\title{Result Table Server}
\usage{
resultTableServer(
  id,
  df,
  colDefsInput,
  addActions = NULL,
  downloadedFileName = NULL
)
}
\arguments{
\item{id}{string, table id must match resultsTableViewer function}

\item{df}{reactive that returns a data frame}

\item{colDefsInput}{named list of reactable::colDefs}

\item{downloadedFileName}{string, desired name of downloaded data file. can use the name from the module that is being used}

\item{addActions}{add a button row selector column to the table to a column called 'actions'.
actions must be a column in df}

}
\value{
shiny module server
}
\description{
Result Table Server
}
