% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-data-viewer.R
\name{resultTableServer}
\alias{resultTableServer}
\title{Result Table Server}
\usage{
resultTableServer(id, df, colDefsInput, downloadedFileName)
}
\arguments{
\item{id}{string, table id must match resultsTableViewer function}

\item{df}{reactive that returns a data frame}

\item{colDefsInput}{named list of reactable::colDefs}

\item{downloadedFileName}{string, desired name of downloaded data file. can use the name from the module that is being used}
}
\value{
shiny module server
}
\description{
Result Table Server
}
