% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-data-viewer.R
\name{resultTableViewer}
\alias{resultTableViewer}
\title{Result Table Viewer}
\usage{
resultTableViewer(id = "result-table")
}
\arguments{
\item{id}{string}
}
\value{
shiny module UI
}
\description{
Result Table Viewer
}
