% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-data-viewer.R
\name{resultTableViewer}
\alias{resultTableViewer}
\title{Result Table Viewer}
\usage{
resultTableViewer(
  id = "result-table",
  downloadedFileName = NULL,
  boxTitle = "Table"
)
}
\arguments{
\item{id}{string}

\item{downloadedFileName}{string, desired name of downloaded data file. can use the name from the module that is being used}

\item{boxTitle}{the title added to the box}
}
\value{
shiny module UI
}
\description{
Result Table Viewer
}
\seealso{
Other {Utils}: 
\code{\link{createCustomColDefList}()},
\code{\link{datasourcesHelperFile}()},
\code{\link{datasourcesServer}()},
\code{\link{datasourcesViewer}()},
\code{\link{makeButtonLabel}()},
\code{\link{resultTableServer}()}
}
\concept{{Utils}}
