% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/components-data-viewer.R
\name{resultTableViewer}
\alias{resultTableViewer}
\title{Result Table Viewer}
\usage{
resultTableViewer(id = "result-table", downloadedFileName)
}
\arguments{
\item{id}{string}

\item{downloadedFileName}{string, desired name of downloaded data file. can use the name from the module that is being used}
}
\value{
shiny module UI
}
\description{
Result Table Viewer
}
