% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-pair.R
\name{selfControlledCohortDetailedColDef}
\alias{selfControlledCohortDetailedColDef}
\title{The column definitions for the detailed pair results table}
\usage{
selfControlledCohortDetailedColDef()
}
\value{
A named list of reactable::colDef
}
\description{
The column definitions for the detailed pair results table
}
\details{
Only the calibrated estimates are shown (together with the analysis
description so the results from different study parameter settings can be
distinguished)
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsViewer}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
