% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-metaExploration.R
\name{selfControlledCohortMetaExplorationColDef}
\alias{selfControlledCohortMetaExplorationColDef}
\title{The column definitions for the meta analytic target-outcome pair table}
\usage{
selfControlledCohortMetaExplorationColDef()
}
\value{
A named list of reactable::colDef
}
\description{
The column definitions for the meta analytic target-outcome pair table
}
\details{
Counts and descriptive statistics are shown for all pairs.  The calibrated
effect estimate columns are returned as NA (displayed as -) for any
evidence synthesis analysis that failed a study diagnostic
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsViewer}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
