% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-metaExploration.R
\name{selfControlledCohortMetaExplorationViewer}
\alias{selfControlledCohortMetaExplorationViewer}
\title{The module viewer for the meta analytic target-outcome pair exploration}
\usage{
selfControlledCohortMetaExplorationViewer(id = "metaExploration")
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the meta analytic exploration module
}
\description{
The module viewer for the meta analytic target-outcome pair exploration
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaSql}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsSql}()},
\code{\link{selfControlledCohortSignalsViewer}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
