% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-main.R
\name{selfControlledCohortServer}
\alias{selfControlledCohortServer}
\title{The module server for exploring self controlled cohort results}
\usage{
selfControlledCohortServer(
  id,
  connectionHandler,
  resultDatabaseSettings = list(port = 1)
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list containing the self controlled cohort
result schema and table prefixes (schema, sccTablePrefix, esTablePrefix,
cgTablePrefix and databaseTable)}
}
\value{
The server for the self controlled cohort module
}
\description{
The module server for exploring self controlled cohort results
}
\details{
The user specifies the id for the module
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortMetaSql}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsSql}()},
\code{\link{selfControlledCohortSignalsViewer}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
