% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-signals.R
\name{selfControlledCohortSignalsServer}
\alias{selfControlledCohortSignalsServer}
\title{The module server for the self controlled cohort signal discovery grid}
\usage{
selfControlledCohortSignalsServer(
  id,
  connectionHandler,
  resultDatabaseSettings,
  selectedPair
)
}
\arguments{
\item{id}{the unique reference id for the module}

\item{connectionHandler}{a connection to the database with the results}

\item{resultDatabaseSettings}{a list with the result schema and table
prefixes}

\item{selectedPair}{a reactiveVal used to communicate the selected pair to
the parent module}
}
\value{
The server for the signal discovery grid
}
\description{
The module server for the self controlled cohort signal discovery grid
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortMetaSql}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsSql}()},
\code{\link{selfControlledCohortSignalsViewer}()},
\code{\link{selfControlledCohortViewer}()}
}
\concept{SelfControlledCohort}
