% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selfControlledCohort-main.R
\name{selfControlledCohortViewer}
\alias{selfControlledCohortViewer}
\title{The module viewer for exploring self controlled cohort results}
\usage{
selfControlledCohortViewer(id = "selfControlledCohort")
}
\arguments{
\item{id}{the unique reference id for the module}
}
\value{
The user interface to the self controlled cohort viewer module
}
\description{
The module viewer for exploring self controlled cohort results
}
\details{
The user specifies the id for the module.  The module provides a signal
discovery view (all exposure-outcome pairs with the number of databases
showing a benefit/risk and the meta analytic estimate) and a pair explorer
for a single exposure-outcome pair (detailed results, forest plot,
systematic error and time at risk / time to outcome boxplots)
}
\seealso{
Other SelfControlledCohort: 
\code{\link{selfControlledCohortCombineResults}()},
\code{\link{selfControlledCohortDetailedColDef}()},
\code{\link{selfControlledCohortDiagnosticsColDef}()},
\code{\link{selfControlledCohortHelperFile}()},
\code{\link{selfControlledCohortMetaExplorationColDef}()},
\code{\link{selfControlledCohortMetaExplorationServer}()},
\code{\link{selfControlledCohortMetaExplorationViewer}()},
\code{\link{selfControlledCohortMetaSql}()},
\code{\link{selfControlledCohortPairServer}()},
\code{\link{selfControlledCohortPairViewer}()},
\code{\link{selfControlledCohortServer}()},
\code{\link{selfControlledCohortSignalsColDef}()},
\code{\link{selfControlledCohortSignalsServer}()},
\code{\link{selfControlledCohortSignalsSql}()},
\code{\link{selfControlledCohortSignalsViewer}()}
}
\concept{SelfControlledCohort}
