% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-timeDistributions.R
\name{timeDistributionsView}
\alias{timeDistributionsView}
\title{timeDistributions view}
\usage{
timeDistributionsView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("imeDistributions") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
