% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cohort-diagnostics-visitContext.R
\name{visitContextView}
\alias{visitContextView}
\title{Visit context module view}
\usage{
visitContextView(id)
}
\arguments{
\item{id}{Namespace Id - use namespaced id ns("vistConext") inside diagnosticsExplorer module}
}
\description{
Use for customizing UI
}
